+++ src/com/almacen/mx/view/aplicaciondepartamentomantenimiento/EntradasMenu.java
@@ -1,0 +1,142 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
+ */
+package com.almacen.mx.view.aplicaciondepartamentomantenimiento;
+
+import aplicaciondepartamentomantenimiento.customUI.CustomHeader;
+import aplicaciondepartamentomantenimiento.customUI.ModeloTabla;
+import aplicaciondepartamentomantenimiento.customUI.ScrollBarCustom;
+import java.awt.Color;
+import java.awt.Font;
+import javax.swing.table.DefaultTableModel;
+import javax.swing.table.JTableHeader;
+
+/**
+ *
+ * @author acost
+ */
+public class EntradasMenu extends javax.swing.JPanel {
+
+    /**
+     * Creates new form EntradasMenu
+     */
+    public EntradasMenu() {
+        initComponents();
+        manualInitComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        panelEntrada = new javax.swing.JPanel();
+        scrollEntradas = new javax.swing.JScrollPane();
+        tablaEntradas = new javax.swing.JTable();
+
+        panelEntrada.setBackground(new java.awt.Color(44, 60, 67));
+        panelEntrada.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                panelEntradaMouseClicked(evt);
+            }
+        });
+
+        scrollEntradas.setBackground(new java.awt.Color(40, 60, 67));
+
+        tablaEntradas.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
+        tablaEntradas.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+
+            },
+            new String [] {
+                "Numero", "Piezas", "Codigo", "Nombre", "Estado", "Proveedor", "Fecha"
+            }
+        ) {
+            boolean[] canEdit = new boolean [] {
+                true, true, true, true, true, true, false
+            };
+
+            public boolean isCellEditable(int rowIndex, int columnIndex) {
+                return canEdit [columnIndex];
+            }
+        });
+        tablaEntradas.setGridColor(new java.awt.Color(40, 60, 67));
+        tablaEntradas.setRowHeight(30);
+        tablaEntradas.setSelectionBackground(new java.awt.Color(0, 204, 102));
+        tablaEntradas.getTableHeader().setResizingAllowed(false);
+        tablaEntradas.getTableHeader().setReorderingAllowed(false);
+        scrollEntradas.setViewportView(tablaEntradas);
+
+        javax.swing.GroupLayout panelEntradaLayout = new javax.swing.GroupLayout(panelEntrada);
+        panelEntrada.setLayout(panelEntradaLayout);
+        panelEntradaLayout.setHorizontalGroup(
+            panelEntradaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelEntradaLayout.createSequentialGroup()
+                .addContainerGap(27, Short.MAX_VALUE)
+                .addComponent(scrollEntradas, javax.swing.GroupLayout.PREFERRED_SIZE, 1065, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(18, 18, 18))
+        );
+        panelEntradaLayout.setVerticalGroup(
+            panelEntradaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(panelEntradaLayout.createSequentialGroup()
+                .addGap(25, 25, 25)
+                .addComponent(scrollEntradas, javax.swing.GroupLayout.PREFERRED_SIZE, 750, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(123, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 1110, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(panelEntrada, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 898, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(panelEntrada, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+    //<editor-fold defaultstate="collapsed" desc="Metodos usados para la interfaz">
+    private void panelEntradaMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_panelEntradaMouseClicked
+        tablaEntradas.clearSelection();
+    }//GEN-LAST:event_panelEntradaMouseClicked
+
+    private void manualInitComponents() {
+        JTableHeader headerEntradas = tablaEntradas.getTableHeader();
+        headerEntradas.setDefaultRenderer(new CustomHeader());
+        tablaEntradas.setTableHeader(headerEntradas);
+        tablaEntradas.getTableHeader().setBackground(new Color(64, 110, 36));
+        tablaEntradas.getTableHeader().setFont(new Font("Corbel", 1, 1));
+        String[] titulos = {"Numero", "Piezas", "Codigo", "Nombre", "Estado", "Proveedor", "Fecha"};
+        Object[][] datos = new Object[0][0];
+        modeloTablaEntrada = new ModeloTabla(datos, titulos, false);
+        tablaEntradas.setModel(modeloTablaEntrada);
+        scrollEntradas.setVerticalScrollBar(new ScrollBarCustom());
+        scrollEntradas.getVerticalScrollBar().setUnitIncrement(30);
+    }
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Variables">
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JPanel panelEntrada;
+    private javax.swing.JScrollPane scrollEntradas;
+    private javax.swing.JTable tablaEntradas;
+    // End of variables declaration//GEN-END:variables
+    DefaultTableModel modeloTablaEntrada;
+    //</editor-fold>
+}
