+++ src/com/almacen/mx/view/aplicaciondepartamentomantenimiento/MenuPrincipal.java
@@ -1,0 +1,529 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.almacen.mx.view.aplicaciondepartamentomantenimiento;
+
+import aplicaciondepartamentomantenimiento.customUI.CustomHeader;
+import aplicaciondepartamentomantenimiento.customUI.ModeloTabla;
+import aplicaciondepartamentomantenimiento.customUI.ScrollBarCustom;
+import java.awt.Color;
+import java.awt.Font;
+import java.awt.Image;
+import java.awt.event.MouseAdapter;
+import java.awt.event.MouseEvent;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import javax.swing.Icon;
+import javax.swing.ImageIcon;
+import javax.swing.JLabel;
+import javax.swing.table.DefaultTableModel;
+import javax.swing.table.JTableHeader;
+
+
+/**
+ *
+ * @author Jerenis Acosta
+ */
+public class MenuPrincipal extends javax.swing.JFrame {
+
+    //Iniciador de la ventana
+    public MenuPrincipal() {
+        initComponents();
+        manualInitComponents();
+        this.setLocationRelativeTo(this);
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        panelPrincipal = new javax.swing.JPanel();
+        barraSuperior = new javax.swing.JPanel();
+        botonMinimizar = new javax.swing.JLabel();
+        botonCerrar = new javax.swing.JLabel();
+        labelMenuActual = new javax.swing.JLabel();
+        labelFecha = new javax.swing.JLabel();
+        scrollPanelMenu = new javax.swing.JScrollPane();
+        panelMenus = new javax.swing.JPanel();
+        entradaMBoton = new javax.swing.JLabel();
+        salidaMboton = new javax.swing.JLabel();
+        stockBoton = new javax.swing.JLabel();
+        cotizacionBoton = new javax.swing.JLabel();
+        seguimientoBoton = new javax.swing.JLabel();
+        cierreMBoton = new javax.swing.JLabel();
+        articulosBoton = new javax.swing.JLabel();
+        ordenBoton = new javax.swing.JLabel();
+        label1 = new javax.swing.JLabel();
+        label2 = new javax.swing.JLabel();
+        label3 = new javax.swing.JLabel();
+        label6 = new javax.swing.JLabel();
+        label4 = new javax.swing.JLabel();
+        panelContenedor = new javax.swing.JTabbedPane();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setTitle("Inventario ");
+        setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        setName("ventana"); // NOI18N
+        setUndecorated(true);
+        setResizable(false);
+        addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                formMouseClicked(evt);
+            }
+        });
+
+        panelPrincipal.setBackground(new java.awt.Color(44, 60, 67));
+        panelPrincipal.setForeground(new java.awt.Color(255, 51, 0));
+        panelPrincipal.setName("panelPrincipal"); // NOI18N
+        panelPrincipal.setPreferredSize(new java.awt.Dimension(1280, 920));
+        panelPrincipal.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        barraSuperior.setBackground(new java.awt.Color(64, 110, 36));
+        barraSuperior.setAutoscrolls(true);
+        barraSuperior.addMouseMotionListener(new java.awt.event.MouseMotionAdapter() {
+            public void mouseDragged(java.awt.event.MouseEvent evt) {
+                barraSuperiorMouseDragged(evt);
+            }
+        });
+        barraSuperior.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mousePressed(java.awt.event.MouseEvent evt) {
+                barraSuperiorMousePressed(evt);
+            }
+        });
+        barraSuperior.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        botonMinimizar.setBackground(new java.awt.Color(241, 46, 101));
+        botonMinimizar.setFont(new java.awt.Font("Tahoma", 0, 24)); // NOI18N
+        botonMinimizar.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        botonMinimizar.setText("-");
+        botonMinimizar.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        botonMinimizar.setOpaque(true);
+        botonMinimizar.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                botonMinimizarMouseClicked(evt);
+            }
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                botonMinimizarMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                botonMinimizarMouseExited(evt);
+            }
+        });
+        barraSuperior.add(botonMinimizar, new org.netbeans.lib.awtextra.AbsoluteConstraints(1200, 10, 30, 20));
+
+        botonCerrar.setBackground(new java.awt.Color(241, 46, 101));
+        botonCerrar.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
+        botonCerrar.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        botonCerrar.setText("X");
+        botonCerrar.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        botonCerrar.setOpaque(true);
+        botonCerrar.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                botonCerrarMouseClicked(evt);
+            }
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                botonCerrarMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                botonCerrarMouseExited(evt);
+            }
+        });
+        barraSuperior.add(botonCerrar, new org.netbeans.lib.awtextra.AbsoluteConstraints(1240, 10, 30, 20));
+        botonCerrar.getAccessibleContext().setAccessibleName("");
+
+        labelMenuActual.setBackground(new java.awt.Color(0, 102, 0));
+        labelMenuActual.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
+        labelMenuActual.setForeground(new java.awt.Color(255, 255, 255));
+        labelMenuActual.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        labelMenuActual.setText("<html><center>TEXTO");
+        labelMenuActual.setAutoscrolls(true);
+        labelMenuActual.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        labelMenuActual.setOpaque(true);
+        labelMenuActual.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
+        barraSuperior.add(labelMenuActual, new org.netbeans.lib.awtextra.AbsoluteConstraints(120, 0, 190, 40));
+        labelMenuActual.getAccessibleContext().setAccessibleName("FECHA");
+
+        labelFecha.setBackground(new java.awt.Color(0, 102, 0));
+        labelFecha.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
+        labelFecha.setForeground(new java.awt.Color(255, 255, 255));
+        labelFecha.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        labelFecha.setText("<html><center>FECHA");
+        labelFecha.setAutoscrolls(true);
+        labelFecha.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        labelFecha.setOpaque(true);
+        labelFecha.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
+        barraSuperior.add(labelFecha, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 130, 40));
+
+        panelPrincipal.add(barraSuperior, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 1280, 40));
+
+        scrollPanelMenu.setBackground(new java.awt.Color(0, 153, 51));
+        scrollPanelMenu.setPreferredSize(new java.awt.Dimension(80, 1000));
+
+        panelMenus.setBackground(new java.awt.Color(119, 168, 37));
+        panelMenus.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        panelMenus.setMinimumSize(new java.awt.Dimension(150, 0));
+        panelMenus.setPreferredSize(new java.awt.Dimension(10, 2430));
+        panelMenus.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        entradaMBoton.setBackground(new java.awt.Color(102, 153, 0));
+        entradaMBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        entradaMBoton.setToolTipText("");
+        entradaMBoton.setVerticalAlignment(javax.swing.SwingConstants.BOTTOM);
+        entradaMBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        entradaMBoton.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
+        entradaMBoton.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                entradaMBotonMouseClicked(evt);
+            }
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                entradaMBotonMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                entradaMBotonMouseExited(evt);
+            }
+        });
+        panelMenus.add(entradaMBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 10, 140, 140));
+        entradaMBoton.getAccessibleContext().setAccessibleName("entrada");
+
+        salidaMboton.setBackground(new java.awt.Color(102, 153, 0));
+        salidaMboton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        salidaMboton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        salidaMboton.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                salidaMbotonMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                salidaMbotonMouseExited(evt);
+            }
+        });
+        panelMenus.add(salidaMboton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 230, 140, 140));
+        salidaMboton.getAccessibleContext().setAccessibleName("salida");
+
+        stockBoton.setBackground(new java.awt.Color(102, 153, 0));
+        stockBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        stockBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        stockBoton.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                stockBotonMouseEntered(evt);
+            }
+        });
+        panelMenus.add(stockBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 450, 140, 140));
+        stockBoton.getAccessibleContext().setAccessibleName("stock");
+
+        cotizacionBoton.setBackground(new java.awt.Color(102, 153, 0));
+        cotizacionBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        cotizacionBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        panelMenus.add(cotizacionBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 670, 140, 140));
+        cotizacionBoton.getAccessibleContext().setAccessibleName("cotizacion");
+
+        seguimientoBoton.setBackground(new java.awt.Color(102, 153, 0));
+        seguimientoBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        seguimientoBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        panelMenus.add(seguimientoBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 890, 140, 140));
+        seguimientoBoton.getAccessibleContext().setAccessibleName("seguimiento");
+
+        cierreMBoton.setBackground(new java.awt.Color(102, 153, 0));
+        cierreMBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        cierreMBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        panelMenus.add(cierreMBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 1110, 140, 140));
+
+        articulosBoton.setBackground(new java.awt.Color(102, 153, 0));
+        articulosBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        articulosBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        panelMenus.add(articulosBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 1330, 140, 140));
+
+        ordenBoton.setBackground(new java.awt.Color(102, 153, 0));
+        ordenBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        ordenBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        panelMenus.add(ordenBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 1550, 140, 140));
+
+        label1.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
+        label1.setForeground(new java.awt.Color(44, 60, 67));
+        label1.setText("<html><center>Entrada de  <p> material");
+        panelMenus.add(label1, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 150, 130, -1));
+
+        label2.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
+        label2.setForeground(new java.awt.Color(44, 60, 67));
+        label2.setText("<html><center>Salida de  <p> material");
+        panelMenus.add(label2, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 370, 110, -1));
+
+        label3.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
+        label3.setForeground(new java.awt.Color(44, 60, 67));
+        label3.setText("<html><center>Stock");
+        panelMenus.add(label3, new org.netbeans.lib.awtextra.AbsoluteConstraints(50, 590, 110, -1));
+
+        label6.setFont(new java.awt.Font("Corbel", 0, 22)); // NOI18N
+        label6.setForeground(new java.awt.Color(44, 60, 67));
+        label6.setText("<html><center>Cotizacion");
+        panelMenus.add(label6, new org.netbeans.lib.awtextra.AbsoluteConstraints(25, 810, 110, -1));
+
+        label4.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
+        label4.setForeground(new java.awt.Color(44, 60, 67));
+        label4.setText("<html><center>Seguimiento");
+        panelMenus.add(label4, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 1030, 180, -1));
+
+        scrollPanelMenu.setViewportView(panelMenus);
+
+        panelPrincipal.add(scrollPanelMenu, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 40, 180, 900));
+
+        panelContenedor.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
+        panelContenedor.setMinimumSize(new java.awt.Dimension(81, 125));
+        panelPrincipal.add(panelContenedor, new org.netbeans.lib.awtextra.AbsoluteConstraints(170, 10, 1110, 930));
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(panelPrincipal, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(panelPrincipal, javax.swing.GroupLayout.DEFAULT_SIZE, 940, Short.MAX_VALUE)
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    //<editor-fold defaultstate="collapsed" desc="Metodos usados para la interfaz">    
+    private void botonCerrarMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonCerrarMouseClicked
+        System.exit(0);
+    }//GEN-LAST:event_botonCerrarMouseClicked
+
+    private void botonCerrarMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonCerrarMouseEntered
+        botonCerrar.setBackground(new Color(241, 200, 200));
+    }//GEN-LAST:event_botonCerrarMouseEntered
+
+    private void botonCerrarMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonCerrarMouseExited
+        botonCerrar.setBackground(new Color(241, 46, 101));
+    }//GEN-LAST:event_botonCerrarMouseExited
+
+    private void entradaMBotonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_entradaMBotonMouseEntered
+        entradaMBoton.setOpaque(true);
+        repaint();
+    }//GEN-LAST:event_entradaMBotonMouseEntered
+
+    private void entradaMBotonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_entradaMBotonMouseExited
+        entradaMBoton.setOpaque(false);
+        repaint();
+    }//GEN-LAST:event_entradaMBotonMouseExited
+
+    private void entradaMBotonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_entradaMBotonMouseClicked
+
+    }//GEN-LAST:event_entradaMBotonMouseClicked
+
+    private void salidaMbotonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_salidaMbotonMouseEntered
+        salidaMboton.setOpaque(true);
+        repaint();
+    }//GEN-LAST:event_salidaMbotonMouseEntered
+
+    private void salidaMbotonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_salidaMbotonMouseExited
+        salidaMboton.setOpaque(false);
+        repaint();
+    }//GEN-LAST:event_salidaMbotonMouseExited
+
+    private void stockBotonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_stockBotonMouseEntered
+
+    }//GEN-LAST:event_stockBotonMouseEntered
+
+    private void barraSuperiorMouseDragged(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_barraSuperiorMouseDragged
+        int x = evt.getXOnScreen();
+        int y = evt.getYOnScreen();
+        this.setLocation(x - xMouse, y - yMouse);
+    }//GEN-LAST:event_barraSuperiorMouseDragged
+
+    private void barraSuperiorMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_barraSuperiorMousePressed
+        xMouse = evt.getX();
+        yMouse = evt.getY();
+    }//GEN-LAST:event_barraSuperiorMousePressed
+
+    private void botonMinimizarMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonMinimizarMouseClicked
+        this.setExtendedState(MenuPrincipal.ICONIFIED);
+    }//GEN-LAST:event_botonMinimizarMouseClicked
+
+    private void botonMinimizarMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonMinimizarMouseEntered
+        botonMinimizar.setBackground(new Color(241, 200, 200));
+    }//GEN-LAST:event_botonMinimizarMouseEntered
+
+    private void botonMinimizarMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonMinimizarMouseExited
+        botonMinimizar.setBackground(new Color(241, 46, 101));
+    }//GEN-LAST:event_botonMinimizarMouseExited
+
+    private void formMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseClicked
+
+    }//GEN-LAST:event_formMouseClicked
+
+    private void botonesMenuMouseEntered(MouseEvent evt) {
+        JLabel boton = (JLabel) evt.getSource();
+        boton.setOpaque(true);
+        repaint();
+    }
+
+    private void botonesMenuMouseExited(MouseEvent evt) {
+        JLabel boton = (JLabel) evt.getSource();
+        boton.setOpaque(false);
+        repaint();
+    }
+
+    private void manualInitComponents() {
+        ponerImagen(entradaMBoton, "src/com/almacen/mx/view/recursos/entradaMLogo.png");
+        ponerImagen(salidaMboton, "src/com/almacen/mx/view/recursos/salidaMlogo.png");
+        ponerImagen(stockBoton, "src/com/almacen/mx/view/recursos/stocklogo.png");
+        ponerImagen(cotizacionBoton, "src/com/almacen/mx/view/recursos/cotizacionlogo.png");
+        ponerImagen(seguimientoBoton, "src/com/almacen/mx/view/recursos/seguimientologo.png");
+        ponerImagen(cierreMBoton, "src/com/almacen/mx/view/recursos/cierrelogo.png");
+        ponerImagen(articulosBoton, "src/com/almacen/mx/view/recursos/articuloslogo.png");
+        ponerImagen(ordenBoton, "src/com/almacen/mx/view/recursos/ordenlogo.png");
+
+        maBotonesMenu = new MouseAdapter() {
+            public void mouseEntered(MouseEvent evt) {
+                botonesMenuMouseEntered(evt);
+            }
+
+            public void mouseExited(MouseEvent evt) {
+                botonesMenuMouseExited(evt);
+            }
+
+            public void mouseClicked(MouseEvent evt) {
+                JLabel e = (JLabel) evt.getSource();
+                if (e.equals(entradaMBoton)) {
+                    panelContenedor.setSelectedIndex(1);
+                    labelMenuActual.setText("Entrada");
+                    return;
+                }
+                if (e.equals(salidaMboton)) {
+                    panelContenedor.setSelectedIndex(2);
+                    labelMenuActual.setText("Salida");
+                    return;
+                }
+                if (e.equals(stockBoton)) {
+                    panelContenedor.setSelectedIndex(3);
+                    labelMenuActual.setText("Stock");
+                    return;
+                }
+                if (e.equals(cotizacionBoton)) {
+                    panelContenedor.setSelectedIndex(4);
+                    labelMenuActual.setText("Cotizacion");
+                    return;
+                }
+                if (e.equals(seguimientoBoton)) {
+                    panelContenedor.setSelectedIndex(5);
+                    labelMenuActual.setText("Seguimiento");
+                    return;
+                }
+            }
+        };
+
+        entradaMBoton.addMouseListener(maBotonesMenu);
+        salidaMboton.addMouseListener(maBotonesMenu);
+        stockBoton.addMouseListener(maBotonesMenu);
+        cotizacionBoton.addMouseListener(maBotonesMenu);
+        seguimientoBoton.addMouseListener(maBotonesMenu);
+
+        scrollPanelMenu.setVerticalScrollBar(new ScrollBarCustom());
+        scrollPanelMenu.getVerticalScrollBar().setUnitIncrement(30);
+
+        labelFecha.setText(formatter.format(date));
+        labelMenuActual.setText("Bienvenido");
+
+        panelContenedor.add(bm);
+        panelContenedor.add(em);
+        panelContenedor.add(sm);
+        panelContenedor.add(stm);
+        panelContenedor.add(cm);
+        panelContenedor.add(sgm);
+
+        panelContenedor.setSelectedIndex(0);
+
+    }
+
+    private void ponerImagen(JLabel lb, String ruta) {
+        ImageIcon imagen = new ImageIcon(ruta);
+        Icon icono = new ImageIcon(imagen.getImage().getScaledInstance(lb.getWidth(), lb.getHeight(), Image.SCALE_SMOOTH));
+        lb.setIcon(icono);
+        this.repaint();
+    }
+
+    //</editor-fold>
+    
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(MenuPrincipal.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(MenuPrincipal.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(MenuPrincipal.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(MenuPrincipal.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+        //</editor-fold>
+
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new MenuPrincipal().setVisible(true);
+            }
+        });
+    }
+
+    //<editor-fold defaultstate="collapsed" desc="Variables">
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JLabel articulosBoton;
+    private javax.swing.JPanel barraSuperior;
+    private javax.swing.JLabel botonCerrar;
+    private javax.swing.JLabel botonMinimizar;
+    private javax.swing.JLabel cierreMBoton;
+    private javax.swing.JLabel cotizacionBoton;
+    private javax.swing.JLabel entradaMBoton;
+    private javax.swing.JLabel label1;
+    private javax.swing.JLabel label2;
+    private javax.swing.JLabel label3;
+    private javax.swing.JLabel label4;
+    private javax.swing.JLabel label6;
+    private javax.swing.JLabel labelFecha;
+    private javax.swing.JLabel labelMenuActual;
+    private javax.swing.JLabel ordenBoton;
+    private javax.swing.JTabbedPane panelContenedor;
+    private javax.swing.JPanel panelMenus;
+    private javax.swing.JPanel panelPrincipal;
+    private javax.swing.JLabel salidaMboton;
+    private javax.swing.JScrollPane scrollPanelMenu;
+    private javax.swing.JLabel seguimientoBoton;
+    private javax.swing.JLabel stockBoton;
+    // End of variables declaration//GEN-END:variables
+    private MouseAdapter maBotonesMenu;
+    private int xMouse;
+    private int yMouse;
+    DefaultTableModel modeloTablaEntrada;
+    Date date = new Date();
+    SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yy");
+    DefaultTableModel modeloTablaSalida;
+    DefaultTableModel modeloTablaStock;
+    JLabel[] botonesMenu = {entradaMBoton, salidaMboton, stockBoton, cotizacionBoton, seguimientoBoton};
+    EntradasMenu em = new EntradasMenu();
+    SalidasMenu sm = new SalidasMenu();
+    StockMenu stm = new StockMenu();
+    CotizacionMenu cm = new CotizacionMenu();
+    SeguimientoMenu sgm = new SeguimientoMenu();
+    BienvenidaMenu bm = new BienvenidaMenu();
+
+    //</editor-fold>
+}
