+++ src/com/almacen/mx/view/aplicaciondepartamentomantenimiento/SalidasMenu.java
@@ -1,0 +1,128 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
+ */
+package com.almacen.mx.view.aplicaciondepartamentomantenimiento;
+
+import aplicaciondepartamentomantenimiento.customUI.CustomHeader;
+import aplicaciondepartamentomantenimiento.customUI.ModeloTabla;
+import java.awt.Color;
+import java.awt.Font;
+import javax.swing.table.DefaultTableModel;
+import javax.swing.table.JTableHeader;
+
+/**
+ *
+ * @author acost
+ */
+public class SalidasMenu extends javax.swing.JPanel {
+
+    /**
+     * Creates new form SalidasMenu
+     */
+    public SalidasMenu() {
+        initComponents();
+        manualInitComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        panelSalida = new javax.swing.JPanel();
+        scrollSalida = new javax.swing.JScrollPane();
+        tablaSalida = new javax.swing.JTable();
+
+        panelSalida.setBackground(new java.awt.Color(44, 60, 67));
+        panelSalida.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                panelSalidaMouseClicked(evt);
+            }
+        });
+
+        tablaSalida.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
+        tablaSalida.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+
+            },
+            new String [] {
+                "Title 1", "Title 2", "Title 3", "Title 4"
+            }
+        ));
+        tablaSalida.setRowHeight(30);
+        tablaSalida.setSelectionBackground(new java.awt.Color(0, 204, 102));
+        tablaSalida.getTableHeader().setResizingAllowed(false);
+        tablaSalida.getTableHeader().setReorderingAllowed(false);
+        scrollSalida.setViewportView(tablaSalida);
+
+        javax.swing.GroupLayout panelSalidaLayout = new javax.swing.GroupLayout(panelSalida);
+        panelSalida.setLayout(panelSalidaLayout);
+        panelSalidaLayout.setHorizontalGroup(
+            panelSalidaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(panelSalidaLayout.createSequentialGroup()
+                .addGap(21, 21, 21)
+                .addComponent(scrollSalida, javax.swing.GroupLayout.DEFAULT_SIZE, 1083, Short.MAX_VALUE)
+                .addContainerGap())
+        );
+        panelSalidaLayout.setVerticalGroup(
+            panelSalidaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(panelSalidaLayout.createSequentialGroup()
+                .addGap(19, 19, 19)
+                .addComponent(scrollSalida, javax.swing.GroupLayout.PREFERRED_SIZE, 750, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(129, Short.MAX_VALUE))
+        );
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 1110, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(panelSalida, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 898, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(panelSalida, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+    //<editor-fold defaultstate="collapsed" desc="Metodos usados para la interfaz">
+    private void panelSalidaMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_panelSalidaMouseClicked
+        tablaSalida.clearSelection();
+    }//GEN-LAST:event_panelSalidaMouseClicked
+
+    private void manualInitComponents(){
+        JTableHeader headerSalidas = tablaSalida.getTableHeader();
+	headerSalidas.setDefaultRenderer(new CustomHeader());
+	tablaSalida.setTableHeader(headerSalidas);
+        String[] titulosSalida = {"Numero","Piezas","Codigo","Nombre","Personal","Lugar","Estado","Fecha"};
+        Object[][] datosSalida = new Object[0][0];
+        modeloTablaSalida = new ModeloTabla(datosSalida,titulosSalida, false);           
+        tablaSalida.getTableHeader().setBackground(new Color(64,110,36));
+        tablaSalida.getTableHeader().setFont(new Font("Corbel", 1, 1));
+        tablaSalida.setModel(modeloTablaSalida);
+    }
+    //</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="Variables">
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JPanel panelSalida;
+    private javax.swing.JScrollPane scrollSalida;
+    private javax.swing.JTable tablaSalida;
+    // End of variables declaration//GEN-END:variables
+    DefaultTableModel modeloTablaSalida;
+    //</editor-fold>
+}
