+++ src/com/almacen/mx/view/aplicaciondepartamentomantenimiento/StockMenu.java
@@ -1,0 +1,682 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
+ */
+package com.almacen.mx.view.aplicaciondepartamentomantenimiento;
+
+import aplicaciondepartamentomantenimiento.customUI.CustomHeader;
+import aplicaciondepartamentomantenimiento.customUI.ModeloTabla;
+import aplicaciondepartamentomantenimiento.customUI.ScrollBarCustom;
+import java.awt.Color;
+import java.awt.Font;
+import java.awt.Image;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import javax.swing.Icon;
+import javax.swing.ImageIcon;
+import javax.swing.JLabel;
+import javax.swing.table.DefaultTableModel;
+import javax.swing.table.JTableHeader;
+
+/**
+ *
+ * @author acost
+ */
+public class StockMenu extends javax.swing.JPanel {
+
+    /**
+     * Creates new form StockMenu
+     */
+    public StockMenu() {
+        initComponents();       
+        manualInitComponents();
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        panelStock = new javax.swing.JPanel();
+        modificarEntradasMenu = new javax.swing.JPanel();
+        eNumeroTextField = new javax.swing.JTextField();
+        jLabel3 = new javax.swing.JLabel();
+        ePiezasTextField = new javax.swing.JTextField();
+        jLabel4 = new javax.swing.JLabel();
+        eCodigoTextField = new javax.swing.JTextField();
+        jLabel5 = new javax.swing.JLabel();
+        eNombreTextField = new javax.swing.JTextField();
+        jLabel6 = new javax.swing.JLabel();
+        eEstadoTextField = new javax.swing.JTextField();
+        jLabel7 = new javax.swing.JLabel();
+        eProveedorTextField = new javax.swing.JTextField();
+        jLabel8 = new javax.swing.JLabel();
+        eFechaTextField = new javax.swing.JTextField();
+        jLabel9 = new javax.swing.JLabel();
+        eFechaButton = new javax.swing.JButton();
+        eGuardarButton = new javax.swing.JButton();
+        eVolverButton = new javax.swing.JButton();
+        jLabel10 = new javax.swing.JLabel();
+        modificarSalidasMenu = new javax.swing.JPanel();
+        sNumeroTextField = new javax.swing.JTextField();
+        jLabel19 = new javax.swing.JLabel();
+        sPiezasTextField = new javax.swing.JTextField();
+        jLabel20 = new javax.swing.JLabel();
+        sCodigoTextField = new javax.swing.JTextField();
+        jLabel21 = new javax.swing.JLabel();
+        sNombreTextField = new javax.swing.JTextField();
+        jLabel22 = new javax.swing.JLabel();
+        sPersonalTextField = new javax.swing.JTextField();
+        jLabel23 = new javax.swing.JLabel();
+        sLugarTextField = new javax.swing.JTextField();
+        jLabel24 = new javax.swing.JLabel();
+        sFechaTextField = new javax.swing.JTextField();
+        jLabel25 = new javax.swing.JLabel();
+        sFechaButton = new javax.swing.JButton();
+        sGuardarButton = new javax.swing.JButton();
+        sVolverButton = new javax.swing.JButton();
+        jLabel26 = new javax.swing.JLabel();
+        jLabel11 = new javax.swing.JLabel();
+        sEstadoTextField = new javax.swing.JTextField();
+        scrollStock = new javax.swing.JScrollPane();
+        tablaStock = new javax.swing.JTable();
+        modificarEntradasButton = new javax.swing.JLabel();
+        msLabel = new javax.swing.JLabel();
+        modificarSalidasButton = new javax.swing.JLabel();
+        meLabel = new javax.swing.JLabel();
+
+        panelStock.setBackground(new java.awt.Color(44, 60, 67));
+        panelStock.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                panelStockMouseClicked(evt);
+            }
+        });
+        panelStock.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        modificarEntradasMenu.setBackground(new java.awt.Color(204, 204, 204));
+
+        eNumeroTextField.setBackground(new java.awt.Color(255, 255, 255));
+
+        jLabel3.setForeground(new java.awt.Color(0, 0, 0));
+        jLabel3.setText("NUMERO:");
+
+        ePiezasTextField.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                ePiezasTextFieldActionPerformed(evt);
+            }
+        });
+
+        jLabel4.setText("PIEZAS:");
+
+        jLabel5.setText("CODIGO:");
+
+        jLabel6.setText("NOMBRE:");
+
+        eEstadoTextField.setBackground(new java.awt.Color(255, 255, 255));
+
+        jLabel7.setText("ESTADO:");
+
+        eProveedorTextField.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                eProveedorTextFieldActionPerformed(evt);
+            }
+        });
+
+        jLabel8.setText("PROVEEDOR:");
+
+        eFechaTextField.setText("DIA/MES/AÑO");
+        eFechaTextField.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                eFechaTextFieldMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                eFechaTextFieldMouseExited(evt);
+            }
+        });
+
+        jLabel9.setText("FECHA:");
+
+        eFechaButton.setText("HOY");
+        eFechaButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                eFechaButtonActionPerformed(evt);
+            }
+        });
+
+        eGuardarButton.setText("GUARDAR");
+
+        eVolverButton.setText("VOLVER");
+        eVolverButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                eVolverButtonActionPerformed(evt);
+            }
+        });
+
+        jLabel10.setText("REGISTRAR NUEVA ENTRADA");
+
+        javax.swing.GroupLayout modificarEntradasMenuLayout = new javax.swing.GroupLayout(modificarEntradasMenu);
+        modificarEntradasMenu.setLayout(modificarEntradasMenuLayout);
+        modificarEntradasMenuLayout.setHorizontalGroup(
+            modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                        .addGap(84, 84, 84)
+                        .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                .addComponent(jLabel6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addComponent(jLabel5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                            .addComponent(jLabel4)))
+                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                        .addGap(80, 80, 80)
+                        .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 64, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addComponent(eVolverButton)
+                        .addGap(158, 158, 158)
+                        .addComponent(eGuardarButton)
+                        .addGap(259, 259, 259))
+                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addComponent(eNombreTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(ePiezasTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(eNumeroTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(eCodigoTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                    .addComponent(jLabel8)
+                                    .addComponent(jLabel9))
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addComponent(eFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addComponent(eProveedorTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(eFechaButton))
+                            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                                .addGap(230, 230, 230)
+                                .addComponent(jLabel7)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addComponent(eEstadoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(0, 57, Short.MAX_VALUE)))
+                        .addGap(58, 58, 58))))
+            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, modificarEntradasMenuLayout.createSequentialGroup()
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addComponent(jLabel10)
+                .addGap(349, 349, 349))
+        );
+        modificarEntradasMenuLayout.setVerticalGroup(
+            modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
+                .addGap(22, 22, 22)
+                .addComponent(jLabel10)
+                .addGap(32, 32, 32)
+                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(eNumeroTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel3))
+                .addGap(32, 32, 32)
+                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel4)
+                    .addComponent(ePiezasTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(eEstadoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel7))
+                .addGap(48, 48, 48)
+                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel5)
+                    .addComponent(eCodigoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel8)
+                    .addComponent(eProveedorTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(33, 33, 33)
+                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel6)
+                    .addComponent(eNombreTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel9)
+                    .addComponent(eFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(eFechaButton))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 105, Short.MAX_VALUE)
+                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(eGuardarButton)
+                    .addComponent(eVolverButton))
+                .addGap(52, 52, 52))
+        );
+
+        panelStock.add(modificarEntradasMenu, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 110, 880, 480));
+
+        modificarSalidasMenu.setBackground(new java.awt.Color(204, 204, 204));
+
+        sNumeroTextField.setBackground(new java.awt.Color(255, 255, 255));
+
+        jLabel19.setForeground(new java.awt.Color(0, 0, 0));
+        jLabel19.setText("NUMERO:");
+
+        sPiezasTextField.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                sPiezasTextFieldActionPerformed(evt);
+            }
+        });
+
+        jLabel20.setText("PIEZAS:");
+
+        jLabel21.setText("CODIGO:");
+
+        jLabel22.setText("NOMBRE:");
+
+        sPersonalTextField.setBackground(new java.awt.Color(255, 255, 255));
+
+        jLabel23.setText("PERSONAL:");
+
+        sLugarTextField.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                sLugarTextFieldActionPerformed(evt);
+            }
+        });
+
+        jLabel24.setText("LUGAR:");
+
+        sFechaTextField.setText("DIA/MES/AÑO");
+        sFechaTextField.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                sFechaTextFieldMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                sFechaTextFieldMouseExited(evt);
+            }
+        });
+
+        jLabel25.setText("FECHA:");
+
+        sFechaButton.setText("HOY");
+        sFechaButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                sFechaButtonActionPerformed(evt);
+            }
+        });
+
+        sGuardarButton.setText("GUARDAR");
+
+        sVolverButton.setText("VOLVER");
+        sVolverButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                sVolverButtonActionPerformed(evt);
+            }
+        });
+
+        jLabel26.setText("REGISTRAR NUEVA SALIDA");
+
+        jLabel11.setText("ESTADO:");
+
+        javax.swing.GroupLayout modificarSalidasMenuLayout = new javax.swing.GroupLayout(modificarSalidasMenu);
+        modificarSalidasMenu.setLayout(modificarSalidasMenuLayout);
+        modificarSalidasMenuLayout.setHorizontalGroup(
+            modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                        .addGap(84, 84, 84)
+                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                .addComponent(jLabel22, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addComponent(jLabel21, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                            .addComponent(jLabel20)))
+                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                        .addGap(80, 80, 80)
+                        .addComponent(jLabel19, javax.swing.GroupLayout.PREFERRED_SIZE, 64, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                                .addComponent(sNombreTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addGap(251, 251, 251)
+                                .addComponent(jLabel25))
+                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                                .addComponent(sVolverButton)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addComponent(jLabel26)
+                                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                                        .addGap(167, 167, 167)
+                                        .addComponent(sGuardarButton)))))
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(sFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(sFechaButton)
+                        .addGap(60, 60, 60))
+                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, modificarSalidasMenuLayout.createSequentialGroup()
+                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                                        .addComponent(sNumeroTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addGap(218, 218, 218)
+                                        .addComponent(jLabel23))
+                                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                                        .addComponent(sPiezasTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(jLabel24)))
+                                .addGap(19, 19, 19))
+                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                                .addComponent(sCodigoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                .addComponent(jLabel11)
+                                .addGap(18, 18, 18)))
+                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                            .addComponent(sPersonalTextField, javax.swing.GroupLayout.DEFAULT_SIZE, 164, Short.MAX_VALUE)
+                            .addComponent(sLugarTextField)
+                            .addComponent(sEstadoTextField))
+                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
+        );
+        modificarSalidasMenuLayout.setVerticalGroup(
+            modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                .addGap(20, 20, 20)
+                .addComponent(jLabel26)
+                .addGap(32, 32, 32)
+                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                        .addComponent(sPersonalTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addComponent(jLabel23))
+                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
+                        .addGap(2, 2, 2)
+                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(sNumeroTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addComponent(jLabel19))))
+                .addGap(32, 32, 32)
+                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel20)
+                    .addComponent(sPiezasTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(sLugarTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel24))
+                .addGap(48, 48, 48)
+                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel21)
+                    .addComponent(sCodigoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel11)
+                    .addComponent(sEstadoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addGap(33, 33, 33)
+                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel22)
+                    .addComponent(sNombreTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel25)
+                    .addComponent(sFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(sFechaButton))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(sGuardarButton)
+                    .addComponent(sVolverButton))
+                .addGap(52, 52, 52))
+        );
+
+        panelStock.add(modificarSalidasMenu, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 110, 880, 480));
+
+        tablaStock.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
+        tablaStock.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+                {},
+                {},
+                {},
+                {},
+                {},
+                {}
+            },
+            new String [] {
+
+            }
+        ));
+        tablaStock.setAutoscrolls(false);
+        tablaStock.setRowHeight(30);
+        tablaStock.setSelectionBackground(new java.awt.Color(0, 204, 102));
+        tablaStock.getTableHeader().setResizingAllowed(false);
+        tablaStock.getTableHeader().setReorderingAllowed(false);
+        scrollStock.setViewportView(tablaStock);
+
+        panelStock.add(scrollStock, new org.netbeans.lib.awtextra.AbsoluteConstraints(31, 30, 1056, 719));
+
+        modificarEntradasButton.setBackground(new java.awt.Color(0, 51, 51));
+        modificarEntradasButton.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                modificarEntradasButtonMouseClicked(evt);
+            }
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                modificarEntradasButtonMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                modificarEntradasButtonMouseExited(evt);
+            }
+        });
+        panelStock.add(modificarEntradasButton, new org.netbeans.lib.awtextra.AbsoluteConstraints(723, 761, 63, 65));
+
+        msLabel.setFont(new java.awt.Font("Corbel", 0, 12)); // NOI18N
+        msLabel.setForeground(new java.awt.Color(255, 255, 255));
+        msLabel.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        msLabel.setText("<HTML><center>MODIFICAR <p> SALIDAS");
+        msLabel.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
+        panelStock.add(msLabel, new org.netbeans.lib.awtextra.AbsoluteConstraints(294, 832, -1, -1));
+
+        modificarSalidasButton.setBackground(new java.awt.Color(0, 51, 51));
+        modificarSalidasButton.addMouseListener(new java.awt.event.MouseAdapter() {
+            public void mouseClicked(java.awt.event.MouseEvent evt) {
+                modificarSalidasButtonMouseClicked(evt);
+            }
+            public void mouseEntered(java.awt.event.MouseEvent evt) {
+                modificarSalidasButtonMouseEntered(evt);
+            }
+            public void mouseExited(java.awt.event.MouseEvent evt) {
+                modificarSalidasButtonMouseExited(evt);
+            }
+        });
+        panelStock.add(modificarSalidasButton, new org.netbeans.lib.awtextra.AbsoluteConstraints(294, 761, 63, 65));
+
+        meLabel.setFont(new java.awt.Font("Corbel", 0, 12)); // NOI18N
+        meLabel.setForeground(new java.awt.Color(255, 255, 255));
+        meLabel.setText("<HTML>MODIFICAR <p> ENTRADAS");
+        panelStock.add(meLabel, new org.netbeans.lib.awtextra.AbsoluteConstraints(723, 832, -1, -1));
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
+        this.setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 1110, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(panelStock, javax.swing.GroupLayout.PREFERRED_SIZE, 1110, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 898, Short.MAX_VALUE)
+            .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(layout.createSequentialGroup()
+                    .addGap(0, 0, Short.MAX_VALUE)
+                    .addComponent(panelStock, javax.swing.GroupLayout.PREFERRED_SIZE, 898, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addGap(0, 0, Short.MAX_VALUE)))
+        );
+    }// </editor-fold>//GEN-END:initComponents
+
+    // <editor-fold defaultstate="collapsed" desc="Metodos usados para la interfaz">
+    private void ePiezasTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_ePiezasTextFieldActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_ePiezasTextFieldActionPerformed
+
+    private void eProveedorTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_eProveedorTextFieldActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_eProveedorTextFieldActionPerformed
+
+    private void eFechaTextFieldMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_eFechaTextFieldMouseEntered
+        if(eFechaTextField.getText().isEmpty() || eFechaTextField.getText().isBlank() || eFechaTextField.getText().equals("DIA/MES/AÑO")){
+            eFechaTextField.setText("");
+        }
+    }//GEN-LAST:event_eFechaTextFieldMouseEntered
+
+    private void eFechaTextFieldMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_eFechaTextFieldMouseExited
+        if(eFechaTextField.getText().isEmpty() || eFechaTextField.getText().isBlank()){
+            eFechaTextField.setText("DIA/MES/AÑO");
+        }
+    }//GEN-LAST:event_eFechaTextFieldMouseExited
+
+    private void eFechaButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_eFechaButtonActionPerformed
+        eFechaTextField.setText(formatter.format(date));
+    }//GEN-LAST:event_eFechaButtonActionPerformed
+
+    private void eVolverButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_eVolverButtonActionPerformed
+        modificarEntradasMenu.setVisible(false);
+        scrollStock.setVisible(true);
+    }//GEN-LAST:event_eVolverButtonActionPerformed
+
+    private void modificarEntradasButtonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarEntradasButtonMouseClicked
+        modificarEntradasMenu.setVisible(true);
+        scrollStock.setVisible(false);
+        modificarSalidasMenu.setVisible(false);
+    }//GEN-LAST:event_modificarEntradasButtonMouseClicked
+
+    private void modificarEntradasButtonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarEntradasButtonMouseEntered
+        modificarEntradasButton.setOpaque(true);
+        repaint();
+    }//GEN-LAST:event_modificarEntradasButtonMouseEntered
+
+    private void modificarEntradasButtonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarEntradasButtonMouseExited
+        modificarEntradasButton.setOpaque(false);
+        repaint();
+    }//GEN-LAST:event_modificarEntradasButtonMouseExited
+
+    private void modificarSalidasButtonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarSalidasButtonMouseClicked
+        modificarSalidasMenu.setVisible(true);
+        scrollStock.setVisible(false);
+        modificarEntradasMenu.setVisible(false);
+    }//GEN-LAST:event_modificarSalidasButtonMouseClicked
+
+    private void modificarSalidasButtonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarSalidasButtonMouseEntered
+        modificarSalidasButton.setOpaque(true);
+        repaint();
+    }//GEN-LAST:event_modificarSalidasButtonMouseEntered
+
+    private void modificarSalidasButtonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarSalidasButtonMouseExited
+        modificarSalidasButton.setOpaque(false);
+        repaint();
+    }//GEN-LAST:event_modificarSalidasButtonMouseExited
+
+    private void sPiezasTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sPiezasTextFieldActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_sPiezasTextFieldActionPerformed
+
+    private void sLugarTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sLugarTextFieldActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_sLugarTextFieldActionPerformed
+
+    private void sFechaTextFieldMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_sFechaTextFieldMouseEntered
+        if(sFechaTextField.getText().isEmpty() || sFechaTextField.getText().isBlank() || sFechaTextField.getText().equals("DIA/MES/AÑO")){
+            sFechaTextField.setText("");
+        }
+    }//GEN-LAST:event_sFechaTextFieldMouseEntered
+
+    private void sFechaTextFieldMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_sFechaTextFieldMouseExited
+        if(sFechaTextField.getText().isEmpty() || sFechaTextField.getText().isBlank()){
+            sFechaTextField.setText("DIA/MES/AÑO");
+        }
+    }//GEN-LAST:event_sFechaTextFieldMouseExited
+
+    private void sFechaButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sFechaButtonActionPerformed
+        sFechaTextField.setText(formatter.format(date));
+    }//GEN-LAST:event_sFechaButtonActionPerformed
+
+    private void sVolverButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sVolverButtonActionPerformed
+        modificarSalidasMenu.setVisible(false);
+        scrollStock.setVisible(true);
+    }//GEN-LAST:event_sVolverButtonActionPerformed
+
+    private void panelStockMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_panelStockMouseClicked
+        tablaStock.clearSelection();
+    }//GEN-LAST:event_panelStockMouseClicked
+    
+    private void manualInitComponents(){
+        JTableHeader headerStock = tablaStock.getTableHeader();
+	headerStock.setDefaultRenderer(new CustomHeader());
+	tablaStock.setTableHeader(headerStock);
+        String[] titulosStock = {"Numero","Piezas","Codigo","Nombre","Modelo","Proveedor","Estado","Observacion","Fecha"};
+        Object[][] datosStock = new Object[0][0];
+        modeloTablaStock = new ModeloTabla(datosStock,titulosStock, false);           
+        tablaStock.getTableHeader().setBackground(new Color(64,110,36));
+        tablaStock.getTableHeader().setFont(new Font("Corbel", 1, 1));
+        tablaStock.setModel(modeloTablaStock);
+        scrollStock.setVerticalScrollBar(new ScrollBarCustom());
+        scrollStock.getVerticalScrollBar().setUnitIncrement(30);
+        
+        modificarEntradasMenu.setVisible(false);
+        modificarSalidasMenu.setVisible(false);
+        modificarSalidasButton.setSize(63, 65);
+        modificarEntradasButton.setSize(63, 65);
+        ponerImagen(modificarSalidasButton, "src/com/almacen/mx/view/recursos/ajustesstock.png");
+        ponerImagen(modificarEntradasButton, "src/com/almacen/mx/view/recursos/ajustesstock.png");
+    }
+    
+    private void ponerImagen(JLabel lb, String ruta) {
+        ImageIcon imagen = new ImageIcon(ruta);
+        Icon icono = new ImageIcon(imagen.getImage().getScaledInstance(lb.getWidth(), lb.getHeight(), Image.SCALE_SMOOTH));
+        lb.setIcon(icono);
+        lb.repaint();
+    }
+    // </editor-fold> 
+    
+    //<editor-fold defaultstate="collapsed" desc="Variables">
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JTextField eCodigoTextField;
+    private javax.swing.JTextField eEstadoTextField;
+    private javax.swing.JButton eFechaButton;
+    private javax.swing.JTextField eFechaTextField;
+    private javax.swing.JButton eGuardarButton;
+    private javax.swing.JTextField eNombreTextField;
+    private javax.swing.JTextField eNumeroTextField;
+    private javax.swing.JTextField ePiezasTextField;
+    private javax.swing.JTextField eProveedorTextField;
+    private javax.swing.JButton eVolverButton;
+    private javax.swing.JLabel jLabel10;
+    private javax.swing.JLabel jLabel11;
+    private javax.swing.JLabel jLabel19;
+    private javax.swing.JLabel jLabel20;
+    private javax.swing.JLabel jLabel21;
+    private javax.swing.JLabel jLabel22;
+    private javax.swing.JLabel jLabel23;
+    private javax.swing.JLabel jLabel24;
+    private javax.swing.JLabel jLabel25;
+    private javax.swing.JLabel jLabel26;
+    private javax.swing.JLabel jLabel3;
+    private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
+    private javax.swing.JLabel jLabel6;
+    private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel jLabel8;
+    private javax.swing.JLabel jLabel9;
+    private javax.swing.JLabel meLabel;
+    private javax.swing.JLabel modificarEntradasButton;
+    private javax.swing.JPanel modificarEntradasMenu;
+    private javax.swing.JLabel modificarSalidasButton;
+    private javax.swing.JPanel modificarSalidasMenu;
+    private javax.swing.JLabel msLabel;
+    private javax.swing.JPanel panelStock;
+    private javax.swing.JTextField sCodigoTextField;
+    private javax.swing.JTextField sEstadoTextField;
+    private javax.swing.JButton sFechaButton;
+    private javax.swing.JTextField sFechaTextField;
+    private javax.swing.JButton sGuardarButton;
+    private javax.swing.JTextField sLugarTextField;
+    private javax.swing.JTextField sNombreTextField;
+    private javax.swing.JTextField sNumeroTextField;
+    private javax.swing.JTextField sPersonalTextField;
+    private javax.swing.JTextField sPiezasTextField;
+    private javax.swing.JButton sVolverButton;
+    private javax.swing.JScrollPane scrollStock;
+    private javax.swing.JTable tablaStock;
+    // End of variables declaration//GEN-END:variables
+    Date date = new Date();
+    SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yy"); 
+    DefaultTableModel modeloTablaStock;
+    //</editor-fold>
+}
