--- src/com/almacen/mx/view/aplicaciondepartamentomantenimiento/Ventana.java
@@ -1,1393 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.almacen.mx.view.aplicaciondepartamentomantenimiento;
-
-import aplicaciondepartamentomantenimiento.customUI.CustomHeader;
-import aplicaciondepartamentomantenimiento.customUI.ModeloTabla;
-import aplicaciondepartamentomantenimiento.customUI.ScrollBarCustom;
-import java.awt.Color;
-import java.awt.Component;
-import java.awt.Dimension;
-import java.awt.Font;
-import java.awt.Image;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.event.MouseAdapter;
-import java.awt.event.MouseEvent;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-import java.util.Objects;
-import javax.swing.Icon;
-import javax.swing.ImageIcon;
-import javax.swing.JButton;
-import javax.swing.JComponent;
-import javax.swing.JLabel;
-import javax.swing.JTable;
-import javax.swing.SwingConstants;
-import javax.swing.table.DefaultTableCellRenderer;
-import javax.swing.table.DefaultTableModel;
-import javax.swing.table.JTableHeader;
-
-
-/**
- *
- * @author Jerenis Acosta
- */
-public class Ventana extends javax.swing.JFrame {
-
-    //Iniciador de la ventana
-    public Ventana() {
-        initComponents();
-        manualInitComponents();
-        this.setLocationRelativeTo(this);       
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        panelPrincipal = new javax.swing.JPanel();
-        barraSuperior = new javax.swing.JPanel();
-        botonMinimizar = new javax.swing.JLabel();
-        botonCerrar = new javax.swing.JLabel();
-        labelMenuActual = new javax.swing.JLabel();
-        labelFecha = new javax.swing.JLabel();
-        scrollPanelMenu = new javax.swing.JScrollPane();
-        panelMenu = new javax.swing.JPanel();
-        entradaMBoton = new javax.swing.JLabel();
-        salidaMboton = new javax.swing.JLabel();
-        stockBoton = new javax.swing.JLabel();
-        cotizacionBoton = new javax.swing.JLabel();
-        seguimientoBoton = new javax.swing.JLabel();
-        cierreMBoton = new javax.swing.JLabel();
-        articulosBoton = new javax.swing.JLabel();
-        ordenBoton = new javax.swing.JLabel();
-        label1 = new javax.swing.JLabel();
-        label2 = new javax.swing.JLabel();
-        label3 = new javax.swing.JLabel();
-        label6 = new javax.swing.JLabel();
-        label4 = new javax.swing.JLabel();
-        panelContenedor = new javax.swing.JTabbedPane();
-        panelEntrada = new javax.swing.JPanel();
-        scrollEntradas = new javax.swing.JScrollPane();
-        tablaEntradas = new javax.swing.JTable();
-        panelSalida = new javax.swing.JPanel();
-        scrollSalida = new javax.swing.JScrollPane();
-        tablaSalida = new javax.swing.JTable();
-        panelStock = new javax.swing.JPanel();
-        modificarEntradasMenu = new javax.swing.JPanel();
-        eNumeroTextField = new javax.swing.JTextField();
-        jLabel3 = new javax.swing.JLabel();
-        ePiezasTextField = new javax.swing.JTextField();
-        jLabel4 = new javax.swing.JLabel();
-        eCodigoTextField = new javax.swing.JTextField();
-        jLabel5 = new javax.swing.JLabel();
-        eNombreTextField = new javax.swing.JTextField();
-        jLabel6 = new javax.swing.JLabel();
-        eEstadoTextField = new javax.swing.JTextField();
-        jLabel7 = new javax.swing.JLabel();
-        eProveedorTextField = new javax.swing.JTextField();
-        jLabel8 = new javax.swing.JLabel();
-        eFechaTextField = new javax.swing.JTextField();
-        jLabel9 = new javax.swing.JLabel();
-        eFechaButton = new javax.swing.JButton();
-        eGuardarButton = new javax.swing.JButton();
-        eVolverButton = new javax.swing.JButton();
-        jLabel10 = new javax.swing.JLabel();
-        scrollStock = new javax.swing.JScrollPane();
-        tablaStock = new javax.swing.JTable();
-        modificarEntradasButton = new javax.swing.JLabel();
-        msLabel = new javax.swing.JLabel();
-        modificarSalidasButton = new javax.swing.JLabel();
-        meLabel = new javax.swing.JLabel();
-        modificarSalidasMenu = new javax.swing.JPanel();
-        sNumeroTextField = new javax.swing.JTextField();
-        jLabel19 = new javax.swing.JLabel();
-        sPiezasTextField = new javax.swing.JTextField();
-        jLabel20 = new javax.swing.JLabel();
-        sCodigoTextField = new javax.swing.JTextField();
-        jLabel21 = new javax.swing.JLabel();
-        sNombreTextField = new javax.swing.JTextField();
-        jLabel22 = new javax.swing.JLabel();
-        sPersonalTextField = new javax.swing.JTextField();
-        jLabel23 = new javax.swing.JLabel();
-        sLugarTextField = new javax.swing.JTextField();
-        jLabel24 = new javax.swing.JLabel();
-        sFechaTextField = new javax.swing.JTextField();
-        jLabel25 = new javax.swing.JLabel();
-        sFechaButton = new javax.swing.JButton();
-        sGuardarButton = new javax.swing.JButton();
-        sVolverButton = new javax.swing.JButton();
-        jLabel26 = new javax.swing.JLabel();
-        jLabel11 = new javax.swing.JLabel();
-        sEstadoTextField = new javax.swing.JTextField();
-        panelCotizacion = new javax.swing.JPanel();
-        labelCotizacion = new javax.swing.JLabel();
-        panelSeguimiento = new javax.swing.JPanel();
-        labelSeguimiento = new javax.swing.JLabel();
-        panelMenuPrincipal = new javax.swing.JPanel();
-        logoITS = new javax.swing.JLabel();
-        textoInstitucion = new javax.swing.JLabel();
-        jLabel1 = new javax.swing.JLabel();
-        jLabel2 = new javax.swing.JLabel();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setTitle("Inventario ");
-        setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
-        setName("ventana"); // NOI18N
-        setUndecorated(true);
-        setResizable(false);
-        addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                formMouseClicked(evt);
-            }
-        });
-
-        panelPrincipal.setBackground(new java.awt.Color(44, 60, 67));
-        panelPrincipal.setForeground(new java.awt.Color(255, 51, 0));
-        panelPrincipal.setName("panelPrincipal"); // NOI18N
-        panelPrincipal.setPreferredSize(new java.awt.Dimension(1280, 920));
-        panelPrincipal.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        barraSuperior.setBackground(new java.awt.Color(64, 110, 36));
-        barraSuperior.setAutoscrolls(true);
-        barraSuperior.addMouseMotionListener(new java.awt.event.MouseMotionAdapter() {
-            public void mouseDragged(java.awt.event.MouseEvent evt) {
-                barraSuperiorMouseDragged(evt);
-            }
-        });
-        barraSuperior.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mousePressed(java.awt.event.MouseEvent evt) {
-                barraSuperiorMousePressed(evt);
-            }
-        });
-        barraSuperior.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        botonMinimizar.setBackground(new java.awt.Color(241, 46, 101));
-        botonMinimizar.setFont(new java.awt.Font("Tahoma", 0, 24)); // NOI18N
-        botonMinimizar.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        botonMinimizar.setText("-");
-        botonMinimizar.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        botonMinimizar.setOpaque(true);
-        botonMinimizar.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                botonMinimizarMouseClicked(evt);
-            }
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                botonMinimizarMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                botonMinimizarMouseExited(evt);
-            }
-        });
-        barraSuperior.add(botonMinimizar, new org.netbeans.lib.awtextra.AbsoluteConstraints(1200, 10, 30, 20));
-
-        botonCerrar.setBackground(new java.awt.Color(241, 46, 101));
-        botonCerrar.setFont(new java.awt.Font("Tahoma", 0, 12)); // NOI18N
-        botonCerrar.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        botonCerrar.setText("X");
-        botonCerrar.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        botonCerrar.setOpaque(true);
-        botonCerrar.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                botonCerrarMouseClicked(evt);
-            }
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                botonCerrarMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                botonCerrarMouseExited(evt);
-            }
-        });
-        barraSuperior.add(botonCerrar, new org.netbeans.lib.awtextra.AbsoluteConstraints(1240, 10, 30, 20));
-        botonCerrar.getAccessibleContext().setAccessibleName("");
-
-        labelMenuActual.setBackground(new java.awt.Color(0, 102, 0));
-        labelMenuActual.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
-        labelMenuActual.setForeground(new java.awt.Color(255, 255, 255));
-        labelMenuActual.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        labelMenuActual.setText("<html><center>TEXTO");
-        labelMenuActual.setAutoscrolls(true);
-        labelMenuActual.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        labelMenuActual.setOpaque(true);
-        labelMenuActual.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
-        barraSuperior.add(labelMenuActual, new org.netbeans.lib.awtextra.AbsoluteConstraints(120, 0, 190, 40));
-        labelMenuActual.getAccessibleContext().setAccessibleName("FECHA");
-
-        labelFecha.setBackground(new java.awt.Color(0, 102, 0));
-        labelFecha.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
-        labelFecha.setForeground(new java.awt.Color(255, 255, 255));
-        labelFecha.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        labelFecha.setText("<html><center>FECHA");
-        labelFecha.setAutoscrolls(true);
-        labelFecha.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        labelFecha.setOpaque(true);
-        labelFecha.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
-        barraSuperior.add(labelFecha, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 130, 40));
-
-        panelPrincipal.add(barraSuperior, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 1280, 40));
-
-        scrollPanelMenu.setBackground(new java.awt.Color(0, 153, 51));
-        scrollPanelMenu.setPreferredSize(new java.awt.Dimension(80, 1000));
-
-        panelMenu.setBackground(new java.awt.Color(119, 168, 37));
-        panelMenu.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
-        panelMenu.setMinimumSize(new java.awt.Dimension(150, 0));
-        panelMenu.setPreferredSize(new java.awt.Dimension(10, 2430));
-        panelMenu.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        entradaMBoton.setBackground(new java.awt.Color(102, 153, 0));
-        entradaMBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        entradaMBoton.setToolTipText("");
-        entradaMBoton.setVerticalAlignment(javax.swing.SwingConstants.BOTTOM);
-        entradaMBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        entradaMBoton.setVerticalTextPosition(javax.swing.SwingConstants.BOTTOM);
-        entradaMBoton.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                entradaMBotonMouseClicked(evt);
-            }
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                entradaMBotonMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                entradaMBotonMouseExited(evt);
-            }
-        });
-        panelMenu.add(entradaMBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 10, 140, 140));
-        entradaMBoton.getAccessibleContext().setAccessibleName("entrada");
-
-        salidaMboton.setBackground(new java.awt.Color(102, 153, 0));
-        salidaMboton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        salidaMboton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        salidaMboton.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                salidaMbotonMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                salidaMbotonMouseExited(evt);
-            }
-        });
-        panelMenu.add(salidaMboton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 230, 140, 140));
-        salidaMboton.getAccessibleContext().setAccessibleName("salida");
-
-        stockBoton.setBackground(new java.awt.Color(102, 153, 0));
-        stockBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        stockBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        stockBoton.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                stockBotonMouseEntered(evt);
-            }
-        });
-        panelMenu.add(stockBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 450, 140, 140));
-        stockBoton.getAccessibleContext().setAccessibleName("stock");
-
-        cotizacionBoton.setBackground(new java.awt.Color(102, 153, 0));
-        cotizacionBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        cotizacionBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        panelMenu.add(cotizacionBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 670, 140, 140));
-        cotizacionBoton.getAccessibleContext().setAccessibleName("cotizacion");
-
-        seguimientoBoton.setBackground(new java.awt.Color(102, 153, 0));
-        seguimientoBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        seguimientoBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        panelMenu.add(seguimientoBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 890, 140, 140));
-        seguimientoBoton.getAccessibleContext().setAccessibleName("seguimiento");
-
-        cierreMBoton.setBackground(new java.awt.Color(102, 153, 0));
-        cierreMBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        cierreMBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        panelMenu.add(cierreMBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 1110, 140, 140));
-
-        articulosBoton.setBackground(new java.awt.Color(102, 153, 0));
-        articulosBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        articulosBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        panelMenu.add(articulosBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 1330, 140, 140));
-
-        ordenBoton.setBackground(new java.awt.Color(102, 153, 0));
-        ordenBoton.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        ordenBoton.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        panelMenu.add(ordenBoton, new org.netbeans.lib.awtextra.AbsoluteConstraints(10, 1550, 140, 140));
-
-        label1.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
-        label1.setForeground(new java.awt.Color(44, 60, 67));
-        label1.setText("<html><center>Entrada de  <p> material");
-        panelMenu.add(label1, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 150, 130, -1));
-
-        label2.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
-        label2.setForeground(new java.awt.Color(44, 60, 67));
-        label2.setText("<html><center>Salida de  <p> material");
-        panelMenu.add(label2, new org.netbeans.lib.awtextra.AbsoluteConstraints(30, 370, 110, -1));
-
-        label3.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
-        label3.setForeground(new java.awt.Color(44, 60, 67));
-        label3.setText("<html><center>Stock");
-        panelMenu.add(label3, new org.netbeans.lib.awtextra.AbsoluteConstraints(50, 590, 110, -1));
-
-        label6.setFont(new java.awt.Font("Corbel", 0, 22)); // NOI18N
-        label6.setForeground(new java.awt.Color(44, 60, 67));
-        label6.setText("<html><center>Cotizacion");
-        panelMenu.add(label6, new org.netbeans.lib.awtextra.AbsoluteConstraints(25, 810, 110, -1));
-
-        label4.setFont(new java.awt.Font("Corbel", 0, 24)); // NOI18N
-        label4.setForeground(new java.awt.Color(44, 60, 67));
-        label4.setText("<html><center>Seguimiento");
-        panelMenu.add(label4, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 1030, 180, -1));
-
-        scrollPanelMenu.setViewportView(panelMenu);
-
-        panelPrincipal.add(scrollPanelMenu, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 40, 180, 900));
-
-        panelContenedor.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
-        panelContenedor.setMinimumSize(new java.awt.Dimension(81, 125));
-
-        panelEntrada.setBackground(new java.awt.Color(44, 60, 67));
-        panelEntrada.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                panelEntradaMouseClicked(evt);
-            }
-        });
-
-        scrollEntradas.setBackground(new java.awt.Color(40, 60, 67));
-
-        tablaEntradas.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
-        tablaEntradas.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-
-            },
-            new String [] {
-                "Numero", "Piezas", "Codigo", "Nombre", "Estado", "Proveedor", "Fecha"
-            }
-        ) {
-            boolean[] canEdit = new boolean [] {
-                true, true, true, true, true, true, false
-            };
-
-            public boolean isCellEditable(int rowIndex, int columnIndex) {
-                return canEdit [columnIndex];
-            }
-        });
-        tablaEntradas.setGridColor(new java.awt.Color(40, 60, 67));
-        tablaEntradas.setRowHeight(30);
-        tablaEntradas.setSelectionBackground(new java.awt.Color(0, 204, 102));
-        tablaEntradas.getTableHeader().setResizingAllowed(false);
-        tablaEntradas.getTableHeader().setReorderingAllowed(false);
-        scrollEntradas.setViewportView(tablaEntradas);
-        if (tablaEntradas.getColumnModel().getColumnCount() > 0) {
-            tablaEntradas.getColumnModel().getColumn(0).setResizable(false);
-            tablaEntradas.getColumnModel().getColumn(1).setResizable(false);
-            tablaEntradas.getColumnModel().getColumn(2).setResizable(false);
-            tablaEntradas.getColumnModel().getColumn(3).setResizable(false);
-            tablaEntradas.getColumnModel().getColumn(4).setResizable(false);
-            tablaEntradas.getColumnModel().getColumn(5).setResizable(false);
-            tablaEntradas.getColumnModel().getColumn(6).setResizable(false);
-        }
-
-        javax.swing.GroupLayout panelEntradaLayout = new javax.swing.GroupLayout(panelEntrada);
-        panelEntrada.setLayout(panelEntradaLayout);
-        panelEntradaLayout.setHorizontalGroup(
-            panelEntradaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelEntradaLayout.createSequentialGroup()
-                .addContainerGap(27, Short.MAX_VALUE)
-                .addComponent(scrollEntradas, javax.swing.GroupLayout.PREFERRED_SIZE, 1065, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18))
-        );
-        panelEntradaLayout.setVerticalGroup(
-            panelEntradaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelEntradaLayout.createSequentialGroup()
-                .addGap(25, 25, 25)
-                .addComponent(scrollEntradas, javax.swing.GroupLayout.PREFERRED_SIZE, 750, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(123, Short.MAX_VALUE))
-        );
-
-        panelContenedor.addTab("tab1", panelEntrada);
-
-        panelSalida.setBackground(new java.awt.Color(44, 60, 67));
-        panelSalida.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                panelSalidaMouseClicked(evt);
-            }
-        });
-
-        tablaSalida.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
-        tablaSalida.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-
-            },
-            new String [] {
-                "Title 1", "Title 2", "Title 3", "Title 4"
-            }
-        ));
-        tablaSalida.setRowHeight(30);
-        tablaSalida.setSelectionBackground(new java.awt.Color(0, 204, 102));
-        tablaSalida.getTableHeader().setResizingAllowed(false);
-        tablaSalida.getTableHeader().setReorderingAllowed(false);
-        scrollSalida.setViewportView(tablaSalida);
-
-        javax.swing.GroupLayout panelSalidaLayout = new javax.swing.GroupLayout(panelSalida);
-        panelSalida.setLayout(panelSalidaLayout);
-        panelSalidaLayout.setHorizontalGroup(
-            panelSalidaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelSalidaLayout.createSequentialGroup()
-                .addGap(21, 21, 21)
-                .addComponent(scrollSalida, javax.swing.GroupLayout.DEFAULT_SIZE, 1083, Short.MAX_VALUE)
-                .addContainerGap())
-        );
-        panelSalidaLayout.setVerticalGroup(
-            panelSalidaLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelSalidaLayout.createSequentialGroup()
-                .addGap(19, 19, 19)
-                .addComponent(scrollSalida, javax.swing.GroupLayout.PREFERRED_SIZE, 750, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(129, Short.MAX_VALUE))
-        );
-
-        panelContenedor.addTab("tab2", panelSalida);
-
-        panelStock.setBackground(new java.awt.Color(44, 60, 67));
-        panelStock.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                panelStockMouseClicked(evt);
-            }
-        });
-        panelStock.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
-
-        modificarEntradasMenu.setBackground(new java.awt.Color(204, 204, 204));
-
-        eNumeroTextField.setBackground(new java.awt.Color(255, 255, 255));
-
-        jLabel3.setForeground(new java.awt.Color(0, 0, 0));
-        jLabel3.setText("NUMERO:");
-
-        ePiezasTextField.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                ePiezasTextFieldActionPerformed(evt);
-            }
-        });
-
-        jLabel4.setText("PIEZAS:");
-
-        jLabel5.setText("CODIGO:");
-
-        jLabel6.setText("NOMBRE:");
-
-        eEstadoTextField.setBackground(new java.awt.Color(255, 255, 255));
-
-        jLabel7.setText("ESTADO:");
-
-        eProveedorTextField.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                eProveedorTextFieldActionPerformed(evt);
-            }
-        });
-
-        jLabel8.setText("PROVEEDOR:");
-
-        eFechaTextField.setText("DIA/MES/AÑO");
-        eFechaTextField.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                eFechaTextFieldMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                eFechaTextFieldMouseExited(evt);
-            }
-        });
-
-        jLabel9.setText("FECHA:");
-
-        eFechaButton.setText("HOY");
-        eFechaButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                eFechaButtonActionPerformed(evt);
-            }
-        });
-
-        eGuardarButton.setText("GUARDAR");
-
-        eVolverButton.setText("VOLVER");
-        eVolverButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                eVolverButtonActionPerformed(evt);
-            }
-        });
-
-        jLabel10.setText("REGISTRAR NUEVA ENTRADA");
-
-        javax.swing.GroupLayout modificarEntradasMenuLayout = new javax.swing.GroupLayout(modificarEntradasMenu);
-        modificarEntradasMenu.setLayout(modificarEntradasMenuLayout);
-        modificarEntradasMenuLayout.setHorizontalGroup(
-            modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                        .addGap(84, 84, 84)
-                        .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                .addComponent(jLabel6, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addComponent(jLabel5, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                            .addComponent(jLabel4)))
-                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                        .addGap(80, 80, 80)
-                        .addComponent(jLabel3, javax.swing.GroupLayout.PREFERRED_SIZE, 64, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(eVolverButton)
-                        .addGap(158, 158, 158)
-                        .addComponent(eGuardarButton)
-                        .addGap(259, 259, 259))
-                    .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addComponent(eNombreTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(ePiezasTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(eNumeroTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(eCodigoTextField, javax.swing.GroupLayout.Alignment.LEADING, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE))
-                        .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                                    .addComponent(jLabel8)
-                                    .addComponent(jLabel9))
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(eFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                    .addComponent(eProveedorTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE))
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(eFechaButton))
-                            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                                .addGap(230, 230, 230)
-                                .addComponent(jLabel7)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addComponent(eEstadoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(0, 57, Short.MAX_VALUE)))
-                        .addGap(58, 58, 58))))
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, modificarEntradasMenuLayout.createSequentialGroup()
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addComponent(jLabel10)
-                .addGap(349, 349, 349))
-        );
-        modificarEntradasMenuLayout.setVerticalGroup(
-            modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(modificarEntradasMenuLayout.createSequentialGroup()
-                .addGap(22, 22, 22)
-                .addComponent(jLabel10)
-                .addGap(32, 32, 32)
-                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(eNumeroTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel3))
-                .addGap(32, 32, 32)
-                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel4)
-                    .addComponent(ePiezasTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(eEstadoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel7))
-                .addGap(48, 48, 48)
-                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel5)
-                    .addComponent(eCodigoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel8)
-                    .addComponent(eProveedorTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(33, 33, 33)
-                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel6)
-                    .addComponent(eNombreTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel9)
-                    .addComponent(eFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(eFechaButton))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 105, Short.MAX_VALUE)
-                .addGroup(modificarEntradasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(eGuardarButton)
-                    .addComponent(eVolverButton))
-                .addGap(52, 52, 52))
-        );
-
-        panelStock.add(modificarEntradasMenu, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 110, 880, 480));
-
-        tablaStock.setFont(new java.awt.Font("Corbel", 1, 14)); // NOI18N
-        tablaStock.setModel(new javax.swing.table.DefaultTableModel(
-            new Object [][] {
-                {},
-                {},
-                {},
-                {},
-                {},
-                {}
-            },
-            new String [] {
-
-            }
-        ));
-        tablaStock.setAutoscrolls(false);
-        tablaStock.setRowHeight(30);
-        tablaStock.setSelectionBackground(new java.awt.Color(0, 204, 102));
-        tablaStock.getTableHeader().setResizingAllowed(false);
-        tablaStock.getTableHeader().setReorderingAllowed(false);
-        scrollStock.setViewportView(tablaStock);
-
-        panelStock.add(scrollStock, new org.netbeans.lib.awtextra.AbsoluteConstraints(31, 30, 1056, 719));
-
-        modificarEntradasButton.setBackground(new java.awt.Color(0, 51, 51));
-        modificarEntradasButton.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                modificarEntradasButtonMouseClicked(evt);
-            }
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                modificarEntradasButtonMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                modificarEntradasButtonMouseExited(evt);
-            }
-        });
-        panelStock.add(modificarEntradasButton, new org.netbeans.lib.awtextra.AbsoluteConstraints(723, 761, 63, 65));
-
-        msLabel.setFont(new java.awt.Font("Corbel", 0, 12)); // NOI18N
-        msLabel.setForeground(new java.awt.Color(255, 255, 255));
-        msLabel.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        msLabel.setText("<HTML><center>MODIFICAR <p> SALIDAS");
-        msLabel.setHorizontalTextPosition(javax.swing.SwingConstants.CENTER);
-        panelStock.add(msLabel, new org.netbeans.lib.awtextra.AbsoluteConstraints(294, 832, -1, -1));
-
-        modificarSalidasButton.setBackground(new java.awt.Color(0, 51, 51));
-        modificarSalidasButton.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseClicked(java.awt.event.MouseEvent evt) {
-                modificarSalidasButtonMouseClicked(evt);
-            }
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                modificarSalidasButtonMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                modificarSalidasButtonMouseExited(evt);
-            }
-        });
-        panelStock.add(modificarSalidasButton, new org.netbeans.lib.awtextra.AbsoluteConstraints(294, 761, 63, 65));
-
-        meLabel.setFont(new java.awt.Font("Corbel", 0, 12)); // NOI18N
-        meLabel.setForeground(new java.awt.Color(255, 255, 255));
-        meLabel.setText("<HTML>MODIFICAR <p> ENTRADAS");
-        panelStock.add(meLabel, new org.netbeans.lib.awtextra.AbsoluteConstraints(723, 832, -1, -1));
-
-        modificarSalidasMenu.setBackground(new java.awt.Color(204, 204, 204));
-
-        sNumeroTextField.setBackground(new java.awt.Color(255, 255, 255));
-
-        jLabel19.setForeground(new java.awt.Color(0, 0, 0));
-        jLabel19.setText("NUMERO:");
-
-        sPiezasTextField.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                sPiezasTextFieldActionPerformed(evt);
-            }
-        });
-
-        jLabel20.setText("PIEZAS:");
-
-        jLabel21.setText("CODIGO:");
-
-        jLabel22.setText("NOMBRE:");
-
-        sPersonalTextField.setBackground(new java.awt.Color(255, 255, 255));
-
-        jLabel23.setText("PERSONAL:");
-
-        sLugarTextField.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                sLugarTextFieldActionPerformed(evt);
-            }
-        });
-
-        jLabel24.setText("LUGAR:");
-
-        sFechaTextField.setText("DIA/MES/AÑO");
-        sFechaTextField.addMouseListener(new java.awt.event.MouseAdapter() {
-            public void mouseEntered(java.awt.event.MouseEvent evt) {
-                sFechaTextFieldMouseEntered(evt);
-            }
-            public void mouseExited(java.awt.event.MouseEvent evt) {
-                sFechaTextFieldMouseExited(evt);
-            }
-        });
-
-        jLabel25.setText("FECHA:");
-
-        sFechaButton.setText("HOY");
-        sFechaButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                sFechaButtonActionPerformed(evt);
-            }
-        });
-
-        sGuardarButton.setText("GUARDAR");
-
-        sVolverButton.setText("VOLVER");
-        sVolverButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                sVolverButtonActionPerformed(evt);
-            }
-        });
-
-        jLabel26.setText("REGISTRAR NUEVA SALIDA");
-
-        jLabel11.setText("ESTADO:");
-
-        javax.swing.GroupLayout modificarSalidasMenuLayout = new javax.swing.GroupLayout(modificarSalidasMenu);
-        modificarSalidasMenu.setLayout(modificarSalidasMenuLayout);
-        modificarSalidasMenuLayout.setHorizontalGroup(
-            modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                        .addGap(84, 84, 84)
-                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                .addComponent(jLabel22, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addComponent(jLabel21, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                            .addComponent(jLabel20)))
-                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                        .addGap(80, 80, 80)
-                        .addComponent(jLabel19, javax.swing.GroupLayout.PREFERRED_SIZE, 64, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                                .addComponent(sNombreTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addGap(251, 251, 251)
-                                .addComponent(jLabel25))
-                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                                .addComponent(sVolverButton)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                                    .addComponent(jLabel26)
-                                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                                        .addGap(167, 167, 167)
-                                        .addComponent(sGuardarButton)))))
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(sFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                        .addComponent(sFechaButton)
-                        .addGap(60, 60, 60))
-                    .addGroup(javax.swing.GroupLayout.Alignment.LEADING, modificarSalidasMenuLayout.createSequentialGroup()
-                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                                        .addComponent(sNumeroTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                        .addGap(218, 218, 218)
-                                        .addComponent(jLabel23))
-                                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                                        .addComponent(sPiezasTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                        .addComponent(jLabel24)))
-                                .addGap(19, 19, 19))
-                            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                                .addComponent(sCodigoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
-                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                                .addComponent(jLabel11)
-                                .addGap(18, 18, 18)))
-                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
-                            .addComponent(sPersonalTextField, javax.swing.GroupLayout.DEFAULT_SIZE, 164, Short.MAX_VALUE)
-                            .addComponent(sLugarTextField)
-                            .addComponent(sEstadoTextField))
-                        .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
-        );
-        modificarSalidasMenuLayout.setVerticalGroup(
-            modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addComponent(jLabel26)
-                .addGap(32, 32, 32)
-                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                        .addComponent(sPersonalTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addComponent(jLabel23))
-                    .addGroup(modificarSalidasMenuLayout.createSequentialGroup()
-                        .addGap(2, 2, 2)
-                        .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                            .addComponent(sNumeroTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                            .addComponent(jLabel19))))
-                .addGap(32, 32, 32)
-                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel20)
-                    .addComponent(sPiezasTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(sLugarTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel24))
-                .addGap(48, 48, 48)
-                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel21)
-                    .addComponent(sCodigoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel11)
-                    .addComponent(sEstadoTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(33, 33, 33)
-                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel22)
-                    .addComponent(sNombreTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel25)
-                    .addComponent(sFechaTextField, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(sFechaButton))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                .addGroup(modificarSalidasMenuLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(sGuardarButton)
-                    .addComponent(sVolverButton))
-                .addGap(52, 52, 52))
-        );
-
-        panelStock.add(modificarSalidasMenu, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 110, 880, 480));
-
-        panelContenedor.addTab("tab3", panelStock);
-
-        panelCotizacion.setBackground(new java.awt.Color(44, 60, 67));
-
-        javax.swing.GroupLayout panelCotizacionLayout = new javax.swing.GroupLayout(panelCotizacion);
-        panelCotizacion.setLayout(panelCotizacionLayout);
-        panelCotizacionLayout.setHorizontalGroup(
-            panelCotizacionLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelCotizacionLayout.createSequentialGroup()
-                .addContainerGap(300, Short.MAX_VALUE)
-                .addComponent(labelCotizacion, javax.swing.GroupLayout.PREFERRED_SIZE, 507, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(298, 298, 298))
-        );
-        panelCotizacionLayout.setVerticalGroup(
-            panelCotizacionLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelCotizacionLayout.createSequentialGroup()
-                .addGap(53, 53, 53)
-                .addComponent(labelCotizacion, javax.swing.GroupLayout.PREFERRED_SIZE, 787, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(59, Short.MAX_VALUE))
-        );
-
-        panelContenedor.addTab("tab6", panelCotizacion);
-
-        panelSeguimiento.setBackground(new java.awt.Color(44, 60, 67));
-
-        javax.swing.GroupLayout panelSeguimientoLayout = new javax.swing.GroupLayout(panelSeguimiento);
-        panelSeguimiento.setLayout(panelSeguimientoLayout);
-        panelSeguimientoLayout.setHorizontalGroup(
-            panelSeguimientoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelSeguimientoLayout.createSequentialGroup()
-                .addContainerGap(305, Short.MAX_VALUE)
-                .addComponent(labelSeguimiento, javax.swing.GroupLayout.PREFERRED_SIZE, 523, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(277, 277, 277))
-        );
-        panelSeguimientoLayout.setVerticalGroup(
-            panelSeguimientoLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelSeguimientoLayout.createSequentialGroup()
-                .addGap(33, 33, 33)
-                .addComponent(labelSeguimiento, javax.swing.GroupLayout.PREFERRED_SIZE, 826, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(40, Short.MAX_VALUE))
-        );
-
-        panelContenedor.addTab("tab7", panelSeguimiento);
-
-        panelMenuPrincipal.setBackground(new java.awt.Color(44, 60, 67));
-
-        textoInstitucion.setFont(new java.awt.Font("Corbel", 1, 48)); // NOI18N
-        textoInstitucion.setForeground(new java.awt.Color(255, 255, 255));
-        textoInstitucion.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        textoInstitucion.setText("INSTITUTO TECNOLOGICO DE SALTILLO");
-        textoInstitucion.setBorder(javax.swing.BorderFactory.createBevelBorder(javax.swing.border.BevelBorder.RAISED));
-
-        jLabel1.setOpaque(true);
-
-        jLabel2.setFont(new java.awt.Font("Corbel", 1, 36)); // NOI18N
-        jLabel2.setForeground(new java.awt.Color(255, 255, 255));
-        jLabel2.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel2.setText("<html><center>Departamento de Mantenimiento <p>Y prevencion");
-
-        javax.swing.GroupLayout panelMenuPrincipalLayout = new javax.swing.GroupLayout(panelMenuPrincipal);
-        panelMenuPrincipal.setLayout(panelMenuPrincipalLayout);
-        panelMenuPrincipalLayout.setHorizontalGroup(
-            panelMenuPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelMenuPrincipalLayout.createSequentialGroup()
-                .addContainerGap(138, Short.MAX_VALUE)
-                .addGroup(panelMenuPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelMenuPrincipalLayout.createSequentialGroup()
-                        .addGroup(panelMenuPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                            .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                            .addComponent(textoInstitucion, javax.swing.GroupLayout.DEFAULT_SIZE, 903, Short.MAX_VALUE))
-                        .addGap(64, 64, 64))
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, panelMenuPrincipalLayout.createSequentialGroup()
-                        .addGroup(panelMenuPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                            .addComponent(jLabel2)
-                            .addComponent(logoITS, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                        .addGap(268, 268, 268))))
-        );
-        panelMenuPrincipalLayout.setVerticalGroup(
-            panelMenuPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(panelMenuPrincipalLayout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(panelMenuPrincipalLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(textoInstitucion, javax.swing.GroupLayout.PREFERRED_SIZE, 96, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 12, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(55, 55, 55)
-                .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(37, 37, 37)
-                .addComponent(logoITS, javax.swing.GroupLayout.PREFERRED_SIZE, 500, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(119, Short.MAX_VALUE))
-        );
-
-        panelContenedor.addTab("tab8", panelMenuPrincipal);
-
-        panelPrincipal.add(panelContenedor, new org.netbeans.lib.awtextra.AbsoluteConstraints(170, 10, 1110, 930));
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(panelPrincipal, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(panelPrincipal, javax.swing.GroupLayout.DEFAULT_SIZE, 940, Short.MAX_VALUE)
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    
-    private void botonCerrarMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonCerrarMouseClicked
-        System.exit(0);
-    }//GEN-LAST:event_botonCerrarMouseClicked
-
-    private void botonCerrarMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonCerrarMouseEntered
-        botonCerrar.setBackground(new Color(241,200,200));
-    }//GEN-LAST:event_botonCerrarMouseEntered
-
-    private void botonCerrarMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonCerrarMouseExited
-        botonCerrar.setBackground(new Color(241,46,101));
-    }//GEN-LAST:event_botonCerrarMouseExited
-
-    private void entradaMBotonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_entradaMBotonMouseEntered
-        entradaMBoton.setOpaque(true);
-        repaint();  
-    }//GEN-LAST:event_entradaMBotonMouseEntered
-
-    private void entradaMBotonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_entradaMBotonMouseExited
-        entradaMBoton.setOpaque(false);
-        repaint();
-    }//GEN-LAST:event_entradaMBotonMouseExited
-
-    private void entradaMBotonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_entradaMBotonMouseClicked
-
-    }//GEN-LAST:event_entradaMBotonMouseClicked
-
-    private void salidaMbotonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_salidaMbotonMouseEntered
-        salidaMboton.setOpaque(true);
-        repaint();
-    }//GEN-LAST:event_salidaMbotonMouseEntered
-
-    private void salidaMbotonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_salidaMbotonMouseExited
-        salidaMboton.setOpaque(false);
-        repaint();
-    }//GEN-LAST:event_salidaMbotonMouseExited
-
-    private void stockBotonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_stockBotonMouseEntered
-        
-    }//GEN-LAST:event_stockBotonMouseEntered
-    
-    private void barraSuperiorMouseDragged(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_barraSuperiorMouseDragged
-        int x = evt.getXOnScreen();
-        int y = evt.getYOnScreen();
-        this.setLocation(x - xMouse, y - yMouse);
-    }//GEN-LAST:event_barraSuperiorMouseDragged
-
-    private void barraSuperiorMousePressed(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_barraSuperiorMousePressed
-        xMouse = evt.getX();
-        yMouse = evt.getY();
-    }//GEN-LAST:event_barraSuperiorMousePressed
-
-    private void botonMinimizarMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonMinimizarMouseClicked
-        this.setExtendedState(Ventana.ICONIFIED);
-    }//GEN-LAST:event_botonMinimizarMouseClicked
-
-    private void botonMinimizarMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonMinimizarMouseEntered
-        botonMinimizar.setBackground(new Color(241,200,200));
-    }//GEN-LAST:event_botonMinimizarMouseEntered
-
-    private void botonMinimizarMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_botonMinimizarMouseExited
-        botonMinimizar.setBackground(new Color(241,46,101));
-    }//GEN-LAST:event_botonMinimizarMouseExited
-
-    private void formMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_formMouseClicked
-
-    }//GEN-LAST:event_formMouseClicked
-
-    private void panelEntradaMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_panelEntradaMouseClicked
-        tablaEntradas.clearSelection();
-    }//GEN-LAST:event_panelEntradaMouseClicked
-
-    private void panelSalidaMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_panelSalidaMouseClicked
-        tablaSalida.clearSelection();
-    }//GEN-LAST:event_panelSalidaMouseClicked
-
-    private void panelStockMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_panelStockMouseClicked
-        tablaStock.clearSelection();
-    }//GEN-LAST:event_panelStockMouseClicked
-
-    private void modificarSalidasButtonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarSalidasButtonMouseEntered
-        modificarSalidasButton.setOpaque(true);
-        repaint();
-    }//GEN-LAST:event_modificarSalidasButtonMouseEntered
-
-    private void modificarSalidasButtonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarSalidasButtonMouseExited
-        modificarSalidasButton.setOpaque(false);
-        repaint();
-    }//GEN-LAST:event_modificarSalidasButtonMouseExited
-
-    private void modificarEntradasButtonMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarEntradasButtonMouseEntered
-        modificarEntradasButton.setOpaque(true);
-        repaint();
-    }//GEN-LAST:event_modificarEntradasButtonMouseEntered
-
-    private void modificarEntradasButtonMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarEntradasButtonMouseExited
-       modificarEntradasButton.setOpaque(false);
-       repaint();
-    }//GEN-LAST:event_modificarEntradasButtonMouseExited
-
-    private void ePiezasTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_ePiezasTextFieldActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_ePiezasTextFieldActionPerformed
-
-    private void eProveedorTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_eProveedorTextFieldActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_eProveedorTextFieldActionPerformed
-
-    private void eFechaButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_eFechaButtonActionPerformed
-        eFechaTextField.setText(formatter.format(date));
-    }//GEN-LAST:event_eFechaButtonActionPerformed
-
-    private void eFechaTextFieldMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_eFechaTextFieldMouseEntered
-        if(eFechaTextField.getText().isEmpty() || eFechaTextField.getText().isBlank() || eFechaTextField.getText().equals("DIA/MES/AÑO")){
-            eFechaTextField.setText("");
-        }
-    }//GEN-LAST:event_eFechaTextFieldMouseEntered
-
-    private void eFechaTextFieldMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_eFechaTextFieldMouseExited
-        if(eFechaTextField.getText().isEmpty() || eFechaTextField.getText().isBlank()){
-            eFechaTextField.setText("DIA/MES/AÑO");
-        }
-    }//GEN-LAST:event_eFechaTextFieldMouseExited
-
-    private void modificarEntradasButtonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarEntradasButtonMouseClicked
-       modificarEntradasMenu.setVisible(true);
-       scrollStock.setVisible(false);
-       modificarSalidasMenu.setVisible(false);
-    }//GEN-LAST:event_modificarEntradasButtonMouseClicked
-
-    private void eVolverButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_eVolverButtonActionPerformed
-        modificarEntradasMenu.setVisible(false);
-       scrollStock.setVisible(true);
-    }//GEN-LAST:event_eVolverButtonActionPerformed
-
-    private void sPiezasTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sPiezasTextFieldActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_sPiezasTextFieldActionPerformed
-
-    private void sLugarTextFieldActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sLugarTextFieldActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_sLugarTextFieldActionPerformed
-
-    private void sFechaTextFieldMouseEntered(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_sFechaTextFieldMouseEntered
-        if(sFechaTextField.getText().isEmpty() || sFechaTextField.getText().isBlank() || sFechaTextField.getText().equals("DIA/MES/AÑO")){
-            sFechaTextField.setText("");
-        }
-    }//GEN-LAST:event_sFechaTextFieldMouseEntered
-
-    private void sFechaTextFieldMouseExited(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_sFechaTextFieldMouseExited
-        if(sFechaTextField.getText().isEmpty() || sFechaTextField.getText().isBlank()){
-            sFechaTextField.setText("DIA/MES/AÑO");
-        }
-    }//GEN-LAST:event_sFechaTextFieldMouseExited
-
-    private void sFechaButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sFechaButtonActionPerformed
-        sFechaTextField.setText(formatter.format(date));
-    }//GEN-LAST:event_sFechaButtonActionPerformed
-
-    private void sVolverButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_sVolverButtonActionPerformed
-       modificarSalidasMenu.setVisible(false);
-       scrollStock.setVisible(true);
-    }//GEN-LAST:event_sVolverButtonActionPerformed
-
-    private void modificarSalidasButtonMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_modificarSalidasButtonMouseClicked
-        modificarSalidasMenu.setVisible(true);
-       scrollStock.setVisible(false);
-       modificarEntradasMenu.setVisible(false);
-    }//GEN-LAST:event_modificarSalidasButtonMouseClicked
-    
-    private void botonesMenuMouseEntered(MouseEvent evt){
-        JLabel boton = (JLabel)evt.getSource();
-        boton.setOpaque(true);
-        repaint();
-    }
-    
-    private void botonesMenuMouseExited(MouseEvent evt){
-        JLabel boton = (JLabel)evt.getSource();
-        boton.setOpaque(false);
-        repaint();
-    }   
-    
-    private void botonesMenuClicked(MouseEvent evt){      
-        for(int i = 0; i < botonesMenu.length; i++){
-            if(evt.getSource().equals(botonesMenu[i])){
-                panelContenedor.setSelectedIndex(i);
-            }
-        }
-    }
-    
-    private void manualInitComponents(){
-        //<editor-fold defaultstate="collapsed" desc="Menu Principal code">
-        ponerImagen(entradaMBoton, "src/com/almacen/mx/view/recursos/entradaMLogo.png");
-        ponerImagen(salidaMboton, "src/com/almacen/mx/view/recursos/salidaMlogo.png");
-        ponerImagen(stockBoton, "src/com/almacen/mx/view/recursos/stocklogo.png");
-        ponerImagen(cotizacionBoton, "src/com/almacen/mx/view/recursos/cotizacionlogo.png");
-        ponerImagen(seguimientoBoton, "src/com/almacen/mx/view/recursos/seguimientologo.png");
-        ponerImagen(cierreMBoton, "src/com/almacen/mx/view/recursos/cierrelogo.png");
-        ponerImagen(articulosBoton, "src/com/almacen/mx/view/recursos/articuloslogo.png");
-        ponerImagen(ordenBoton, "src/com/almacen/mx/view/recursos/ordenlogo.png");
-        ponerImagen(logoITS, "src/com/almacen/mx/view/recursos/logotec.png");
-        ponerImagen(modificarEntradasButton, "src/com/almacen/mx/view/recursos/ajustesstock.png");
-        ponerImagen(modificarSalidasButton, "src/com/almacen/mx/view/recursos/ajustesstock.png");
-                        
-        maBotonesMenu = new MouseAdapter() {
-            public void mouseEntered(MouseEvent evt) {
-                botonesMenuMouseEntered(evt);
-            }
-            public void mouseExited(MouseEvent evt) {
-                botonesMenuMouseExited(evt);
-            }          
-            public void mouseClicked(MouseEvent evt){
-                JLabel e = (JLabel)evt.getSource();
-                if(e.equals(entradaMBoton)){
-                    panelContenedor.setSelectedIndex(0);
-                    labelMenuActual.setText("Entrada");
-                    return;
-                }
-                if(e.equals(salidaMboton)){
-                    panelContenedor.setSelectedIndex(1);
-                    labelMenuActual.setText("Salida");
-                    return;
-                }
-                if(e.equals(stockBoton)){
-                    panelContenedor.setSelectedIndex(2);
-                    labelMenuActual.setText("Stock");
-                    return;
-                }
-                if(e.equals(cotizacionBoton)){
-                    panelContenedor.setSelectedIndex(3);
-                    labelMenuActual.setText("Cotizacion");
-                    return;
-                }
-                if(e.equals(seguimientoBoton)){
-                    panelContenedor.setSelectedIndex(4);
-                    labelMenuActual.setText("Seguimiento");
-                    return;
-                }
-            }
-        };
-        
-        entradaMBoton.addMouseListener(maBotonesMenu);
-        salidaMboton.addMouseListener(maBotonesMenu);
-        stockBoton.addMouseListener(maBotonesMenu);
-        cotizacionBoton.addMouseListener(maBotonesMenu);
-        seguimientoBoton.addMouseListener(maBotonesMenu);
-        
-        scrollPanelMenu.setVerticalScrollBar(new ScrollBarCustom());
-        scrollPanelMenu.getVerticalScrollBar().setUnitIncrement(30);
-        
-        labelFecha.setText(formatter.format(date));
-        labelMenuActual.setText("Bienvenido");
-        panelContenedor.setSelectedIndex(5);
-        //</editor-fold>
-        
-        // <editor-fold defaultstate="collapsed" desc="Menu Entradas Code">              
-        JTableHeader headerEntradas = tablaEntradas.getTableHeader();
-        headerEntradas.setDefaultRenderer(new CustomHeader());
-        tablaEntradas.setTableHeader(headerEntradas);
-        
-        tablaEntradas.getTableHeader().setBackground(new Color(64,110,36));
-        tablaEntradas.getTableHeader().setFont(new Font("Corbel", 1, 1));
-        String[] titulos = {"Numero","Piezas","Codigo","Nombre","Estado","Proveedor","Fecha"};
-        Object[][] datos = new Object[0][0];
-        modeloTablaEntrada = new ModeloTabla(datos,titulos, false);
-        tablaEntradas.setModel(modeloTablaEntrada);
-        scrollEntradas.setVerticalScrollBar(new ScrollBarCustom());
-        scrollEntradas.getVerticalScrollBar().setUnitIncrement(30);
-        //</editor-fold>
-        
-        //<editor-fold defaultstate="collapsed" desc="Menu Salidas Code">
-        
-
-        JTableHeader headerSalidas = tablaSalida.getTableHeader();
-	headerSalidas.setDefaultRenderer(new CustomHeader());
-	tablaSalida.setTableHeader(headerSalidas);
-        String[] titulosSalida = {"Numero","Piezas","Codigo","Nombre","Personal","Lugar","Estado","Fecha"};
-        Object[][] datosSalida = new Object[0][0];
-        modeloTablaSalida = new ModeloTabla(datosSalida,titulosSalida, false);           
-        tablaSalida.getTableHeader().setBackground(new Color(64,110,36));
-        tablaSalida.getTableHeader().setFont(new Font("Corbel", 1, 1));
-        tablaSalida.setModel(modeloTablaSalida);
-        //</editor-fold>
-        
-        //<editor-fold defaultstate="collapsed" desc="Menu Stock Code">
-        JTableHeader headerStock = tablaStock.getTableHeader();
-	headerStock.setDefaultRenderer(new CustomHeader());
-	tablaStock.setTableHeader(headerStock);
-        String[] titulosStock = {"Numero","Piezas","Codigo","Nombre","Modelo","Proveedor","Estado","Observacion","Fecha"};
-        Object[][] datosStock = new Object[0][0];
-        modeloTablaStock = new ModeloTabla(datosStock,titulosStock, false);           
-        tablaStock.getTableHeader().setBackground(new Color(64,110,36));
-        tablaStock.getTableHeader().setFont(new Font("Corbel", 1, 1));
-        tablaStock.setModel(modeloTablaStock);
-        scrollStock.setVerticalScrollBar(new ScrollBarCustom());
-        scrollStock.getVerticalScrollBar().setUnitIncrement(30);
-        
-        modificarEntradasMenu.setVisible(false);
-        modificarSalidasMenu.setVisible(false);
-        //</editor-fold>
-        }
-    private void llevarAMenu(int index){
-        panelContenedor.setSelectedIndex(index);
-    }
-    private void ponerImagen(JLabel lb, String ruta){
-       ImageIcon imagen = new ImageIcon(ruta);
-       Icon icono = new ImageIcon(imagen.getImage().getScaledInstance(lb.getWidth(), lb.getHeight(), Image.SCALE_SMOOTH));
-       lb.setIcon(icono);
-       this.repaint();
-    }
-    
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(Ventana.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(Ventana.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(Ventana.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(Ventana.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new Ventana().setVisible(true);
-            }
-        });
-    }
-//<editor-fold defaultstate="collapsed" desc="Variables">
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JLabel articulosBoton;
-    private javax.swing.JPanel barraSuperior;
-    private javax.swing.JLabel botonCerrar;
-    private javax.swing.JLabel botonMinimizar;
-    private javax.swing.JLabel cierreMBoton;
-    private javax.swing.JLabel cotizacionBoton;
-    private javax.swing.JTextField eCodigoTextField;
-    private javax.swing.JTextField eEstadoTextField;
-    private javax.swing.JButton eFechaButton;
-    private javax.swing.JTextField eFechaTextField;
-    private javax.swing.JButton eGuardarButton;
-    private javax.swing.JTextField eNombreTextField;
-    private javax.swing.JTextField eNumeroTextField;
-    private javax.swing.JTextField ePiezasTextField;
-    private javax.swing.JTextField eProveedorTextField;
-    private javax.swing.JButton eVolverButton;
-    private javax.swing.JLabel entradaMBoton;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JLabel jLabel10;
-    private javax.swing.JLabel jLabel11;
-    private javax.swing.JLabel jLabel19;
-    private javax.swing.JLabel jLabel2;
-    private javax.swing.JLabel jLabel20;
-    private javax.swing.JLabel jLabel21;
-    private javax.swing.JLabel jLabel22;
-    private javax.swing.JLabel jLabel23;
-    private javax.swing.JLabel jLabel24;
-    private javax.swing.JLabel jLabel25;
-    private javax.swing.JLabel jLabel26;
-    private javax.swing.JLabel jLabel3;
-    private javax.swing.JLabel jLabel4;
-    private javax.swing.JLabel jLabel5;
-    private javax.swing.JLabel jLabel6;
-    private javax.swing.JLabel jLabel7;
-    private javax.swing.JLabel jLabel8;
-    private javax.swing.JLabel jLabel9;
-    private javax.swing.JLabel label1;
-    private javax.swing.JLabel label2;
-    private javax.swing.JLabel label3;
-    private javax.swing.JLabel label4;
-    private javax.swing.JLabel label6;
-    private javax.swing.JLabel labelCotizacion;
-    private javax.swing.JLabel labelFecha;
-    private javax.swing.JLabel labelMenuActual;
-    private javax.swing.JLabel labelSeguimiento;
-    private javax.swing.JLabel logoITS;
-    private javax.swing.JLabel meLabel;
-    private javax.swing.JLabel modificarEntradasButton;
-    private javax.swing.JPanel modificarEntradasMenu;
-    private javax.swing.JLabel modificarSalidasButton;
-    private javax.swing.JPanel modificarSalidasMenu;
-    private javax.swing.JLabel msLabel;
-    private javax.swing.JLabel ordenBoton;
-    private javax.swing.JTabbedPane panelContenedor;
-    private javax.swing.JPanel panelCotizacion;
-    private javax.swing.JPanel panelEntrada;
-    private javax.swing.JPanel panelMenu;
-    private javax.swing.JPanel panelMenuPrincipal;
-    private javax.swing.JPanel panelPrincipal;
-    private javax.swing.JPanel panelSalida;
-    private javax.swing.JPanel panelSeguimiento;
-    private javax.swing.JPanel panelStock;
-    private javax.swing.JTextField sCodigoTextField;
-    private javax.swing.JTextField sEstadoTextField;
-    private javax.swing.JButton sFechaButton;
-    private javax.swing.JTextField sFechaTextField;
-    private javax.swing.JButton sGuardarButton;
-    private javax.swing.JTextField sLugarTextField;
-    private javax.swing.JTextField sNombreTextField;
-    private javax.swing.JTextField sNumeroTextField;
-    private javax.swing.JTextField sPersonalTextField;
-    private javax.swing.JTextField sPiezasTextField;
-    private javax.swing.JButton sVolverButton;
-    private javax.swing.JLabel salidaMboton;
-    private javax.swing.JScrollPane scrollEntradas;
-    private javax.swing.JScrollPane scrollPanelMenu;
-    private javax.swing.JScrollPane scrollSalida;
-    private javax.swing.JScrollPane scrollStock;
-    private javax.swing.JLabel seguimientoBoton;
-    private javax.swing.JLabel stockBoton;
-    private javax.swing.JTable tablaEntradas;
-    private javax.swing.JTable tablaSalida;
-    private javax.swing.JTable tablaStock;
-    private javax.swing.JLabel textoInstitucion;
-    // End of variables declaration//GEN-END:variables
-    private MouseAdapter maBotonesMenu;
-    private int xMouse;
-    private int yMouse;
-    DefaultTableModel modeloTablaEntrada;
-    Date date = new Date();
-    SimpleDateFormat formatter = new SimpleDateFormat("dd/MM/yy"); 
-    DefaultTableModel modeloTablaSalida;
-    DefaultTableModel modeloTablaStock;
-    JLabel[] botonesMenu = {entradaMBoton, salidaMboton, stockBoton, cotizacionBoton, seguimientoBoton};
-    //</editor-fold>
-      
-   
-    //</editor-fold>
-}
